refacto: translations part & donations in BTC button
