+++ src/contexts/translations.ts
@@ -1,0 +1,1601 @@
+export const translations = {
+  fr: {
+    // Navigation
+    'nav.home': 'Accueil',
+    'nav.board': 'Gouvernance',
+    'nav.association': 'Association',
+    'nav.branches': 'Antennes',
+    'nav.events': 'Événements',
+    'nav.publications': 'Publications',
+    'nav.donate': 'Nous soutenir',
+
+    // Hero
+    'hero.title': "Le premier réseau",
+    'hero.title2': 'étudiant',
+    'hero.title3': 'Tech',
+    'hero.title4': "d'Europe",
+    'hero.subtitle': "Le réseau physique du numérique : crypto-actifs, IA et cybersécurité",
+    'hero.join': "Rejoindre l'association",
+    'hero.branches': 'Voir nos antennes',
+
+    // Stats
+    'stats.title': 'KRYPTOSPHERE® EN QUELQUES CHIFFRES',
+    'stats.subtitle': "Le réseau étudiant de référence sur les cryptomonnaies et sur la Tech depuis 2017",
+    'stats.students': 'Étudiants',
+    'stats.universities': 'Universités',
+    'stats.events': 'Événements',
+    'stats.branches': 'Villes',
+
+    // Sponsors
+    'sponsors.title': 'Ils nous soutiennent',
+    'sponsors.sponsors': 'Sponsors',
+    'sponsors.partners': 'Partenaires',
+    'sponsors.universities': 'Universités',
+    'sponsors.media': 'Ils parlent de nous',
+
+    // Publications
+    'publications.title': 'Nos dernières publications',
+    'publications.subtitle': 'Découvrez nos dernières analyses et recherches sur les blockchains et les crypto-actifs',
+    'publications.readMore': 'Lire la suite',
+    'publications.viewAll': 'Tout voir',
+
+    // Pages
+    'board.title': 'Notre Bureau',
+    'board.subtitle': "Découvrez les bénévoles qui font battre KRYPTOSPHERE® au quotidien. Des profils diversifiés, des équipes complémentaires, unis par leur confiance et passions.",
+    'board.teamTag': 'Ils font vivre KRYPTOSPHERE® depuis 2017',
+    'board.year': 'Année :',
+    'board.member1.name': 'Raphaël PAUTARD',
+    'board.member1.role': 'Président',
+    'board.member1.description': 'Family Officer Spécialisé en crypto & cofondateur de theintern.fun',
+    'board.member1.school': 'IAE Lyon - KRYPTOSPHERE® FRANCE',
+    'board.member2.name': 'Antoine MANIVAL',
+    'board.member2.role': 'Vice-Président',
+    'board.member2.description': 'Growth & Marketing @NFT Paris, Partnerships @Ledger.',
+    'board.member2.school': 'INSEEC - KRYPTOSPHERE® FRANCE',
+    'board.member3.name': 'Mathieu LARUELLE',
+    'board.member3.role': 'Trésorier',
+    'board.member3.description': 'Blockchain & Software Engineer @Cometh & Louis.finance, Ex Trésorier KRYPTOSPHERE® DeVinci',
+    'board.member3.school': 'ESILV - KRYPTOSPHERE® FRANCE',
+    'board.member4.name': 'Louis PIALLAT',
+    'board.member4.role': 'Secrétaire',
+    'board.member4.description': 'Ex VP KRYPTOSPHERE® ESSEC. Etudiant finance @ESSEC et Maths @Sorbonne',
+    'board.member4.school': 'ESSEC - KRYPTOSPHERE® FRANCE',
+    'board.member5.name': 'Thibault LANGLOIS-BERTHELOT',
+    'board.member5.role': 'Président & Fondateur',
+    'board.member5.description': 'Élève avocat et Docteur en droit',
+    'board.member5.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member6.name': 'William DERENNE',
+    'board.member6.role': 'Administrateur & Membre Fondateur',
+    'board.member6.description': "Professeur de solidity spécialiste d'Ethereum",
+    'board.member6.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member7.name': 'Serge PETROFF',
+    'board.member7.role': 'Président',
+    'board.member7.description': "Consultant spécialiste de la finance décentralisée",
+    'board.member7.school': 'KRYPTOSPHERE® RENNES',
+    'board.member8.name': 'Guillaume GIRARD',
+    'board.member8.role': 'Administrateur & Membre Fondateur',
+    'board.member8.description': "Consultant spécialiste du mining de bitcoins @UTXO Management",
+    'board.member8.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member9.name': 'Othmane EL MEJJADE',
+    'board.member9.role': 'Trésorier Délégué',
+    'board.member9.description': "Consultant et spécialiste en marketing international",
+    'board.member9.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member10.name': 'Fanis MICHALAKIS',
+    'board.member10.role': 'Président & Fondateur',
+    'board.member10.description': "Développeur spécialiste de Bitcoin & Lightning @LN MARKETS",
+    'board.member10.school': 'KRYPTOSPHERE® CENTRALE MARSEILLE',
+    'board.member11.name': 'Titouan GALPIN',
+    'board.member11.role': 'Vice-Président délégué',
+    'board.member11.description': "Fondateur @Luphy.io",
+    'board.member11.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member12.name': 'Julie MARTIN',
+    'board.member12.role': 'Responsable communication',
+    'board.member12.description': "Consultante Product Management & IA",
+    'board.member12.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member13.name': 'Manon SERAUT',
+    'board.member13.role': 'Responsable formations',
+    'board.member13.description': "Responsable ressources humaines @ARCHIMED",
+    'board.member13.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member14.name': 'Harmonie LEFORT',
+    'board.member14.role': 'Responsable événementiel',
+    'board.member14.description': "Relationship Manager @CACEIS",
+    'board.member14.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member15.name': 'Marie CECCHINI',
+    'board.member15.role': 'Responsable partenariats',
+    'board.member15.description': "Gérante de portefeuille en gestion privée @Amplegest.com",
+    'board.member15.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member16.name': 'Aymane TABET',
+    'board.member16.role': 'Responsable R&D',
+    'board.member16.description': "Consultant en Risk Advisory @DELOITTE",
+    'board.member16.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member17.name': 'Karolina GORNA',
+    'board.member17.role': 'Présidente déléguée',
+    'board.member17.description': "Ingénieure doctorante @LEDGER",
+    'board.member17.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member18.name': 'Pierre LOURDELET',
+    'board.member18.role': 'Vice-Président délégué',
+    'board.member18.description': "Data Scientist @Mirakl",
+    'board.member18.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member19.name': 'Mathias VALLET',
+    'board.member19.role': 'Président délégué',
+    'board.member19.description': "Co-fondateur et Président de RUFIJI Capital",
+    'board.member19.school': 'Telecom Sud Paris',
+    'board.member20.name': 'Gabriel BEGAZO',
+    'board.member20.role': 'Trésorier délégué',
+    'board.member20.description': "Co-fondateur et Président de DUSA LABS",
+    'board.member20.school': 'INSA Lyon',
+    'board.member21.name': 'Maxime DIENGER',
+    'board.member21.role': 'Président délégué',
+    'board.member21.description': "Dev Advocate @Ripple",
+    'board.member21.school': 'EMLV',
+    'board.member22.name': 'Léo JOUANNY',
+    'board.member22.role': 'Vice-Président délégué',
+    'board.member22.description': "Founder and CEO @Dusa Labs",
+    'board.member22.school': 'EMLV',
+    'board.member23.name': 'Louis FARCOT',
+    'board.member23.role': 'Trésorier délégué',
+    'board.member23.description': "Chercheur spécialiste des blockchains",
+    'board.member23.school': 'EMLV',
+    'board.member24.name': 'Mathias BOUGON',
+    'board.member24.role': 'Président délégué',
+    'board.member24.description': "Associate Business Developer @HubSpot",
+    'board.member24.school': 'EMLV',
+    'board.member25.name': 'Rayan THIBAULT',
+    'board.member25.role': 'Vice-Président délégué',
+    'board.member25.description': "Ingénieur Projets - Direction Technique @Agap2",
+    'board.member25.school': 'EMLV',
+    'board.member26.name': 'Gildas SOEIRO',
+    'board.member26.role': 'Secrétaire sélégué',
+    'board.member26.description': "Program Manager KRYPTOSPHERE® Accelerator",
+    'board.member26.school': 'EMLV',
+    'board.member27.name': 'Thibault LANGLOIS-BERTHELOT',
+    'board.member27.role': 'Représentant légal & Fondateur',
+    'board.member27.description': "Représentant légal accompagnant l'association dans ses aspects complexes.",
+    'board.member27.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member28.name': 'Choi JANGHYUN',
+    'board.member28.role': 'Responsable IoT',
+    'board.member28.description': "Ph.D Student @KAIST",
+    'board.member28.school': 'KRYPTOSPHERE® IMTBS-TSP',
+    'board.member29.name': 'Brice BONNET',
+    'board.member29.role': 'Responsable Formation, Rédaction / Analyse',
+    'board.member29.description': 'Auditeur Interne spécialiste Banque et Finance',
+    'board.member29.school': 'KRYPTOSPHERE® RENNES SB',
+    'board.member30.name': 'Charles-André GOICHOT',
+    'board.member30.role': 'Trésorier délégué',
+    'board.member30.description': 'Développeur informatique',
+    'board.member30.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member31.name': 'Jules FOA',
+    'board.member31.role': 'Président & Fondateur',
+    'board.member31.description': 'Consultant Web3',
+    'board.member31.school': 'KRYPTOSPHERE® ACCELERATOR',
+    'board.member32.name': 'Maksym ANDRUSHKIV',
+    'board.member32.role': 'Responsable Communication',
+    'board.member32.description': 'Digital Transformation @Staizen Europe',
+    'board.member32.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member33.name': 'Maxence LAISNE',
+    'board.member33.role': 'Secrétaire général délégué',
+    'board.member33.description': 'Tech, Product, Analytics @AMAZON',
+    'board.member33.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member34.name': 'Aymeric LABEDAN',
+    'board.member34.role': 'Responsable des Partenariats',
+    'board.member34.description': 'CTO Advisory Consultant @Wavestone',
+    'board.member34.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member35.name': 'Morgan RICHAUD',
+    'board.member35.role': 'Head of Research',
+    'board.member35.description': 'Marketing & Design Specialist @Deskoin',
+    'board.member35.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member36.name': 'Hadrien JAUBERT',
+    'board.member36.role': 'Président',
+    'board.member36.description': 'Développeur @GOOGLE',
+    'board.member36.school': 'KRYPTOSPHERE® CENTRALE MARSEILLE',
+    'board.member37.name': 'Benoît PREAU',
+    'board.member37.role': 'Head of Writing',
+    'board.member37.description': 'Consultant Energy Transition @Capgemini',
+    'board.member37.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member38.name': 'Emile BREMOND',
+    'board.member38.role': 'Head of Writing & Communication',
+    'board.member38.description': 'Rédacteur @CRYPTOAST',
+    'board.member38.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member39.name': 'Mathis GONTIER DELAUNAY',
+    'board.member39.role': 'Secrétaire Général délégué',
+    'board.member39.description': 'Co-fondateur et le chef de la recherche @Morpho',
+    'board.member39.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member40.name': 'Paul FRAMBOT',
+    'board.member40.role': 'Membre actif',
+    'board.member40.description': 'Fondateur @Morpho',
+    'board.member40.school': 'KRYPTOSPHERE® IMTBS-TSP',
+    'board.member41.name': 'Louis DOIDY',
+    'board.member41.role': 'Responsable Formation',
+    'board.member41.description': 'Data Scientist @LVMH',
+    'board.member41.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member42.name': 'Nandy Bâ',
+    'board.member42.role': 'Head of Research & DeFI',
+    'board.member42.description': 'Formateur et Advisor Web3',
+    'board.member42.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member43.name': 'Nour RAMMAL',
+    'board.member43.role': 'Présidente',
+    'board.member43.description': 'Étudiante en Info et Maths @Institut Polytechnique',
+    'board.member43.school': 'KRYPTOSPHERE® IMTBS-TSP',
+    'board.member44.name': 'Andy COUSINEAU',
+    'board.member44.role': 'Président & Fondateur',
+    'board.member44.description': 'Ingénieur Blockchain @SAHAR',
+    'board.member44.school': 'KRYPTOSPHERE® ESEO ANGERS',
+    'board.member45.name': 'Nasreddine GRIHMA',
+    'board.member45.role': 'Responsable R&D',
+    'board.member45.description': 'Consultant blockchain @IBM',
+    'board.member45.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member46.name': 'Manon COIFFARD',
+    'board.member46.role': 'Responsable événementiel',
+    'board.member46.description': 'Consultante ERP @mc2i',
+    'board.member46.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member47.name': 'Quentin BAILLY',
+    'board.member47.role': 'Membre actif',
+    'board.member47.description': 'M&A @UBS',
+    'board.member47.school': 'KRYPTOSPHERE® LYON',
+    'board.member48.name': 'Pierre-Louis HABOURY',
+    'board.member48.role': 'Head of blockchain programming',
+    'board.member48.description': 'Co-Founder & CTO @Dusa Labs',
+    'board.member48.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member49.name': 'Lilyan BASTIEN',
+    'board.member49.role': 'Membre actif',
+    'board.member49.description': 'Blockchain engineer @RA2 TECH',
+    'board.member49.school': 'KRYPTOSPHERE® DEVINCI',
+    'board.member50.name': 'Pierre FONTAINE',
+    'board.member50.role': 'Secrétaire Général',
+    'board.member50.description': 'Business Analyst @Accenture',
+    'board.member50.school': 'KRYPTOSPHERE® IMTBS-TSP',
+    'board.member51.name': 'Sara FEORE',
+    'board.member51.role': 'Présidente',
+    'board.member51.description': 'Co-fondatrice @CompasSup',
+    'board.member51.school': 'KRYPTOSPHERE® Science Po',
+    'board.member52.name': 'Paul WOLF',
+    'board.member52.role': 'Membre actif',
+    'board.member52.description': 'Responsable Crypto @Mastercard',
+    'board.member52.school': 'KRYPTOSPHERE® ESSEC',
+    'board.member53.name': 'Pierre-Antoine ARSAGUET',
+    'board.member53.role': 'Head of Artificial Intelligence',
+    'board.member53.description': 'Quantitative Researcher @QRT',
+    'board.member53.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member54.name': 'Imen SACI',
+    'board.member54.role': 'Responsable des Partenariats',
+    'board.member54.description': 'Blockchain & Innovation Consultant @IBM',
+    'board.member54.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member55.name': 'Thomas BENETREAU',
+    'board.member55.role': 'Responsable événementiel',
+    'board.member55.description': 'Consultant M&A @Wavestone',
+    'board.member55.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member56.name': 'Clément BIDAN',
+    'board.member56.role': 'Membre actif',
+    'board.member56.description': 'Software Engineer @Datadog',
+    'board.member56.school': 'KRYPTOSPHERE® CENTRALE MARSEILLE',
+    'board.member57.name': 'Jérémie PONTIS',
+    'board.member57.role': 'Trésorier',
+    'board.member57.description': 'Consultant Sénior @Talan Consulting',
+    'board.member57.school': 'KRYPTOSPHERE® IMTBS-TSP Paris',
+    'board.member58.name': 'Lery PHELIPPEAU',
+    'board.member58.role': 'Vice-Président',
+    'board.member58.description': 'Marketing @CoinJar',
+    'board.member58.school': 'KRYPTOSPHERE® LONDON',
+    'board.member59.name': 'Lery PHELIPPEAU',
+    'board.member59.role': 'Vice-Président & Co-fondateur',
+    'board.member59.description': 'Marketing @CoinJar',
+    'board.member59.school': 'KRYPTOSPHERE® LONDON',
+    'board.member60.name': 'Thibaut GRAVELLE-VIVIEN',
+    'board.member60.role': 'Président & Fondateur',
+    'board.member60.description': 'Growth Associate @Merkle Science',
+    'board.member60.school': 'KRYPTOSPHERE® LONDON',
+    'board.member61.name': 'Jean-Baptiste BAROU',
+    'board.member61.role': 'Responsable Ambassadeurs / Alumni',
+    'board.member61.description': 'Outsourcing Analyst @Euronext',
+    'board.member61.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member62.name': 'Mark-Killian ZINENBERG',
+    'board.member62.role': 'DeFi Lead',
+    'board.member62.description': 'Quantitative Developer Intern @Natixis',
+    'board.member62.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member63.name': 'Salim LAKHAL',
+    'board.member63.role': 'Président',
+    'board.member63.description': 'Etudiant @Télécom SudParis',
+    'board.member63.school': 'KRYPTOSPHERE® IMTBS-TSP Paris',
+    'board.member64.name': 'Maxime LACROIX',
+    'board.member64.role': 'Tech Advisor',
+    'board.member64.description': 'Team Lead & Platform Engineer @InstaDeep',
+    'board.member64.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member65.name': 'Emmanuel MARQUETTE',
+    'board.member65.role': 'Président & Fondateur',
+    'board.member65.description': 'Strategy Manager @Lombard Odier',
+    'board.member65.school': 'KRYPTOSPHERE® IMTBS-TSP Paris',
+
+    'board.former.president': 'Ancien Président(e)',
+    'board.former.description': "Ancien président(e) qui a mené(e) notre réseau d'associations vers de nouveaux horizons.",
+    'board.former.description2023': 'Leadership exceptionnel durant l\'année 2023.',
+
+    'board.polesTitle': 'Pôles',
+    'board.poles.partnership.title': 'Pôle Partenariat',
+    'board.poles.partnership.description': 'Le pôle partenariat a pour rôle l\'acquisition et l\'entretien de relations externes à l\'association avec ses partenaires et sponsors. Ses responsables sont les porte-paroles de l\'association et contribuent à sa visibilité dans l\'espace public.',
+    'board.poles.partnership.team.member1.name': 'Jules-Grégoire Lafond',
+    'board.poles.partnership.team.member1.role': 'Responsable',
+    'board.poles.partnership.team.member1.description': 'Co-Fondateur & Ex-Président de l\'antenne KRYPTOSPHERE® NEOMA. Sales at HSBC',
+    'board.poles.partnership.team.member2.name': 'Tevfik Erdal',
+    'board.poles.partnership.team.member2.role': 'Co-responsable',
+    'board.poles.partnership.team.member2.description': 'Ex-Vice Président de l\'antenne KRYPTOSPHERE® NEOMA. Customer Relationship at UBS',
+    'board.poles.partnership.team.member3.name': 'Rayan Kaabache',
+    'board.poles.partnership.team.member3.role': 'Co-Responsable',
+    'board.poles.partnership.team.member3.description': "Co-Fondateur & Ex-Président de l'antenne KRYPTOSPHERE® Jussieu. Étudiant en cybersécurité à Sorbonne Université.",
+    'board.poles.communication.title': 'Pôle Communication & Marketing',
+    'board.poles.communication.description': [
+        'Le pôle Communication & Marketing incarne l’image de KRYPTOSPHERE® et assure sa visibilité auprès du grand public comme acteur clé du Web3. À travers des stratégies créatives et maîtrisées — réseaux sociaux, branding, création de contenu, visuels, campagnes digitales ou partenariats médias — il valorise nos événements, nos initiatives et l’engagement de nos membres pour faire rayonner l’identité de KRYPTOSPHERE® dans l’écosystème Tech européen.',
+    ],
+    'board.poles.communication.team.member1.name': 'Alexandre MANGIN',
+    'board.poles.communication.team.member1.role': 'Responsable',
+    'board.poles.communication.team.member1.description': 'VP KRYPTOSPHERE® iaelyon, Marketing manager at iExec, Community Manager at The Big Whale',
+    'board.poles.communication.team.member2.name': 'Alexandra AH-SOUNE',
+    'board.poles.communication.team.member2.role': 'Co-responsable',
+    'board.poles.communication.team.member2.description': 'Ex-Responsable Communication de l\'antenne KRYPTOSPHERE® ESSEC. Private Banking at Société Générale, student in Finance',
+
+    'association.title': 'L\'Association',
+    'association.subtitle': 'Depuis septembre 2017, KRYPTOSPHERE® s\'est imposé comme une référence européenne de l\'écosystème crypto étudiant. Découvrez notre histoire, nos valeurs et notre vision pour l\'avenir.',
+    'association.historyTag': 'Notre histoire et notre mission',
+    'association.mission': 'Notre Mission',
+    'association.missionText': 'Démocratiser l\'accès aux crypto-actifs et former la prochaine génération aux nouvelles technologies numériques.',
+    'association.values': 'Nos Valeurs',
+    'association.valuesText': 'Confiance, partage de connaissances, collaboration et excellence académique au service de notre réseau de bénévoles.',
+    'association.vision': 'Notre Vision',
+    'association.visionText': "Faire le pont entre l'innovation sur l'IA, la blockchain, l'informatique quantique et l'enseignement supérieur en Europe.",
+    'association.journey': 'Notre Histoire',
+
+    'branches.title': 'Nos Antennes',
+    'branches.subtitle': 'Découvrez nos antennes en Europe et rejoignez la communauté KRYPTOSPHERE® la plus proche de vous.',
+    'branches.discover': 'Découvrir l\'antenne',
+    'branches.localOffice': 'Bureau local',
+    'branches.presenceTag': 'Notre présence en Europe',
+    'branches.parisDescription': 'Nos antennes parisiennes, au cœur de l\'écosystème crypto français.',
+    'branches.parisDescription2': "Découvrez KRYPTOSPHERE® DEVINCI, l'une de nos antennes parisiennes les plus actives !",
+    'branches.parisDescription3': "Découvrez KRYPTOSPHERE® IMTBS-TSP, l'une de nos plus anciennes antennes, parisienne, parmi les plus actives !",
+    'branches.nantesDescription': 'L\'antenne nantaise, dynamique et innovante dans l\'ouest de la France.',
+    'branches.parisDescription4': "L\'antenne de l'ESCP créée en avril 2022, aussi prestigieuse que compétente depuis lors. Spécialisation : crypto, IA.",
+    'branches.londonDescription': 'Notre première antenne internationale, au cœur de la City londonienne.',
+    'branches.lyon': "Créée en février 2024, KS IAE LYON rayonne par l'organisation annuelle de notre évènement phare : le KryptoTour !",
+    'branches.lyonDescription': "Créée en février 2024, KS IAE LYON rayonne par l'organisation annuelle de notre évènement phare : le KryptoTour !",
+    'branches.monsDescription': 'L\'antenne de Mons, au plus près des institutions européennes de Bruxelles.',
+    'branches.backToList': 'Retour aux antennes',
+    'branches.bureauYear': 'Bureau {city} 2025',
+    'branches.contact': 'Nous contacter',
+    'branches.gallery': 'Galerie',
+    'branches.location': 'Localisation',
+    'branches.locationMap': 'Carte de localisation',
+
+    'events.title': 'Événements',
+    'events.subtitle': 'Rejoignez nos événements exclusifs et plongez dans l\'univers blockchain. Workshops, conférences et networking avec les leaders de l\'écosystème crypto.',
+    'events.agendaTag': 'Agenda des événements',
+    'events.lumaButton': 'Voir tous nos événements sur Lu.ma',
+    'events.upcoming': 'Événements à venir',
+    'events.past': 'Événements passés',
+
+    'publications.page.title': 'Publications',
+    'publications.page.subtitle': "Découvrez nos analyses, guides et rapports au sujet des crypto-actifs, de la blockchain ou d'autres technologies. Des ressources créées par et pour la communauté étudiante.",
+    'publications.resourcesTag': 'Ressources et contenus éducatifs',
+    'publications.all': 'Toutes',
+    'publications.kryptopaper': 'KryptoPaper',
+    'publications.articles': 'Articles',
+    'publications.reports': 'Rapports',
+    'publications.studies': 'Études',
+    'publications.read': 'Lire',
+    'publications.pdf': 'PDF',
+    'publications.newsletter.title': 'Restez informés',
+    'publications.newsletter.subtitle': 'Recevez nos dernières publications directement dans votre boîte mail.',
+    'publications.newsletter.placeholder': 'Votre email',
+    'publications.newsletter.subscribe': 'S\'abonner',
+
+    'donation.title': 'Faire une donation',
+    'donation.subtitle': 'Votre soutien nous aide à démocratiser l\'accès aux technologies blockchain et à former la prochaine génération d\'innovateurs européens.',
+    'donation.supportTag': 'Soutenez notre vision et nos activités',
+    'donation.impact': 'Votre impact',
+    'donation.training': 'Former des étudiants',
+    'donation.trainingDesc': 'Financement de workshops, hackathons et formations blockchain',
+    'donation.development': 'Développer nos antennes',
+    'donation.developmentDesc': 'Expansion vers de nouveaux établissements et pays',
+    'donation.events': 'Financer des événements',
+    'donation.eventsDesc': 'Organisation de conférences et meetups gratuits',
+    'donation.choose': 'Faire un don par CB',
+    'donation.choose2': 'Faire un don en Bitcoin',
+    'donation.amount': 'Montant du don',
+    'donation.customAmount': 'Montant personnalisé',
+    'donation.paymentMethod': 'Méthode de paiement',
+    'donation.card': 'Carte bancaire',
+    'donation.crypto': 'Cryptomonnaie',
+    'donation.summary': 'Récapitulatif',
+    'donation.amountLabel': 'Montant du don:',
+    'donation.info': 'Votre don sera utilisé pour soutenir nos activités éducatives et le développement de notre communauté.',
+    'donation.finalize': 'Finaliser la donation',
+    'donation.security': 'Vos dons sont sécurisés et utilisés exclusivement pour les activités de l\'association.',
+    'donation.transparency': 'Transparence financière',
+    'donation.transparencyCRYPTO': 'Faire un don en cryptommonaie',
+    'donation.transparencyBTC': 'Notre noeud Bitcoin',
+    'donation.transparencyText': 'Nous publions une partie de nos informations financières pour assurer une transparence sur nos activités.',
+    'donation.transparencyTextBTC': 'Nous maintenons de façon bénévole notre propre noeud Bitcoin depuis plusieurs années : soutenez-nous ⚡',
+    'donation.transparencyTextCRYPTO': 'Nous acceptons directement les dons en BTC ou en ETH depuis 2021.',
+    'donation.reports': 'Consulter nos données officielles',
+    'donation.reportsBTCNODE': 'Découvrez notre noeud Bitcoin en temps réel',
+    'donation.reportsBTC': 'Donner en BTC (via TOR Browser)',
+    'donation.reportsETH': 'Donner en ETH',
+
+    // Antenne Details
+    'antenne.team': 'Équipe',
+    'antenne.president': 'Président',
+    'antenne.vicePresident': 'Vice-Président',
+    'antenne.secretary': 'Secrétaire',
+    'antenne.treasurer': 'Trésorier',
+    'antenne.contact': 'Contact',
+    'antenne.gallery': 'Galerie',
+    'antenne.location': 'Localisation',
+    'antenne.email': 'Email',
+    'antenne.phone': 'Téléphone',
+    'antenne.address': 'Adresse',
+    'antenne.notFound': 'Antenne non trouvée',
+
+    // Countries
+    'countries.france': 'France',
+    'countries.uk': 'Royaume-Uni',
+    'countries.belgium': 'Belgique',
+
+    // Cities
+    'cities.paris': 'Paris',
+    'cities.nantes': 'Nantes',
+    'cities.lyon': 'Lyon',
+    'cities.london': 'London',
+    'cities.mons': 'Mons',
+    'cities.reims': 'Reims',
+
+    // Team roles
+    'roles.president': 'Président',
+    'roles.vicePresident': 'Vice-Président',
+    'roles.secretary': 'Secrétaire',
+    'roles.treasurer': 'Trésorier',
+    'roles.presidente': 'Présidente',
+    'roles.vicePresidente': 'Vice-Présidente',
+    'roles.tresoriere': 'Trésorière',
+
+    // Footer Legal
+    'footer.legal.disclaimer': 'Avertissement',
+    'footer.legal.privacy': 'Politique de confidentialité',
+    'footer.legal.terms': 'Conditions générales d\'utilisation (CGU)',
+    'footer.legal.cookies': 'Politique de cookies',
+    'footer.legal.legal': 'Mentions légales',
+    'footer.description': 'L\'association étudiante de référence dans la Tech. Rejoignez-nous pour explorer les technologies de demain.',
+    'footer.navigation': 'Navigation',
+    'footer.copyright': '2017 — 2025 | KRYPTOSPHERE® FRANCE | Tous droits réservés.',
+
+    // Contact
+    'contact.email': 'Email',
+    'contact.phone': 'Téléphone',
+    'contact.address': 'Adresse',
+
+    // Timeline years
+    'timeline.2017.title': 'Lancement de KRYPTOSPHERE®',
+    'timeline.2017.description': 'Fondation de notre réseau par 4 étudiants passionnés par Bitcoin',
+    'timeline.2017.achievements': ['Première antenne à KEDGE Marseille', '10 membres fondateurs du réseau', 'Premiers articles sur Bitcoin'],
+
+    'timeline.2018.title': 'Expansion nationale et création de KRYPTOSPHERE® FRANCE',
+    'timeline.2018.description': 'Ouverture des premières antennes en région parisienne et partenariats avec plusieurs grandes écoles et universités',
+    'timeline.2018.achievements': ['4 antennes actives', '50+ membres', 'Premier hackathon blockchain'],
+
+    'timeline.2019.title': "Développement et diversification sur l'IA",
+    'timeline.2019.description': "Création de nouvelles antennes pour accompagner le développement de notre réseau et diversification sur l'IA",
+    'timeline.2019.achievements': ['6 antennes actives', '100+ membres', 'Premier KryptoPaper'],
+
+    'timeline.2020.title': 'Déploiement de notre stratégie de communication',
+    'timeline.2020.description': 'Adaptation aux nouveaux modes de communication',
+    'timeline.2020.achievements': ['Événements virtuels & multiples publications', '10+ antennes actives', '200+ membres'],
+
+    'timeline.2021.title': 'Diversification sur la cybersécurité',
+    'timeline.2021.description': "Ouverture des premières antennes au sein d'écoles d'informatique",
+    'timeline.2021.achievements': ['Nouveaux sponsors', 'Expansion continue & formation interne', '300+ membres'],
+
+    'timeline.2022.title': 'Innovation pédagogique',
+    'timeline.2022.description': 'Lancement de programmes de formations internes avancées sur la blockchain et les crypto-actifs',
+    'timeline.2022.achievements': ['Formations internes labellisées', '15+ antennes', '400+ membres'],
+
+    'timeline.2023.title': 'Consolidation de notre réseau',
+    'timeline.2023.description': "Reconnaissance comme référence en France dans l'enseignement supérieur",
+    'timeline.2023.achievements': ['500+ membres', '20+ antennes', '+7 villes'],
+
+    'timeline.2024.title': 'Internationalisation à Londres et en Belgique',
+    'timeline.2024.description': 'Création de deux associations KRYPTOSPHERE® à Londres et en Belgique',
+    'timeline.2024.achievements': ['Lancement à Londres', 'Lancement en Belgique', '700+ membres'],
+
+    'timeline.2025.title': 'Nouveaux horizons',
+    'timeline.2025.description': "Cibler de nouvelles technologies émergentes et consolider notre position de leader dans l'enseignement supérieur européen",
+    'timeline.2025.achievements': ['Nouveaux sponsors', 'Ordinateurs quantiques', '800+ membres', '3 pays'],
+
+    // Publication items
+    'publication.kryptopaper22.title': 'KryptoPaper Édition 22',
+    'publication.kryptopaper22.date': 'MARS 2022',
+    'publication.kryptopaper22.description': 'Analyse des tendances crypto 2022 et perspectives d\'avenir',
+    'publication.kryptopaper22.readTime': '15 min',
+
+    'publication.defi.title': 'KryptoPaper Édition 2021',
+    'publication.defi.date': 'MARS 2021',
+    'publication.defi.description': 'Introduction complète aux crypto-actifs',
+    'publication.defi.readTime': '10 min',
+
+    'publication.report2024.title': "Rapport d'impact 2024",
+    'publication.report2024.date': 'FEVRIER 2024',
+    'publication.report2024.description': "Bilan chiffrée de l'impact de notre réseau pour le marché des crypto-actifs",
+    'publication.report2024.readTime': '15 min',
+
+    'publication.kryptopaper24.title': 'KryptoPaper Édition 22',
+    'publication.kryptopaper24.date': 'AVRIL 2022',
+    'publication.kryptopaper24.description': 'Actualité, DeFi, Cryptoactifs, Articles de fond',
+    'publication.kryptopaper24.readTime': '18 min',
+
+    'publication.research.title': 'Thèse de doctorat',
+    'publication.research.date': 'AOUT 2023',
+    'publication.research.description': 'La blockchain au regard du droit et de l\'identité',
+    'publication.research.readTime': '60+ min',
+
+    'publication.smartcontracts.title': 'Articles archivés',
+    'publication.smartcontracts.date': '2018 - 2024',
+    'publication.smartcontracts.description': 'Retrouvez les archives de nos anciens articles pour suivre de multiples évolutions technologiques',
+    'publication.smartcontracts.readTime': '+60 min',
+
+    // General
+    'language': 'fr',
+
+// Antenne Hult Business School (London)
+    'antennes.london.name': 'Hult',
+    'antennes.london.description': "KRYPTOSPHERE® LONDON incarne notre expansion européenne, reliant finance traditionnelle et univers crypto.",
+    'antennes.london.bureau.1.name': 'Ray MOUAWAD',
+    'antennes.london.bureau.1.role': 'Président',
+    'antennes.london.bureau.2.name': 'Yanis BOUKHARI',
+    'antennes.london.bureau.2.role': 'Vice-président',
+    'antennes.london.bureau.3.name': 'Emiel VAN RENTERGHEM',
+    'antennes.london.bureau.3.role': 'Trésorier',
+    'antennes.london.bureau.4.name': 'Amine MELLITI',
+    'antennes.london.bureau.4.role': 'Secrétaire général',
+
+    // Antenne Dauphine (Paris)
+    'antennes.dauphine.name': 'Dauphine',
+    'antennes.dauphine.description': "Parce que l'avenir se construit dès aujourd'hui, l'antenne dauphinoise de KRYPTOSPHERE® initie les esprits curieux aux technologies qui transforment notre monde.",
+    'antennes.dauphine.bureau.1.name': 'Ray MOUAWAD',
+    'antennes.dauphine.bureau.1.role': 'Président',
+    'antennes.dauphine.bureau.2.name': 'Yanis BOUKHARI',
+    'antennes.dauphine.bureau.2.role': 'Vice-président',
+    'antennes.dauphine.bureau.3.name': 'Emiel VAN RENTERGHEM',
+    'antennes.dauphine.bureau.3.role': 'Trésorier',
+    'antennes.dauphine.bureau.4.name': 'Amine MELLITI',
+    'antennes.dauphine.bureau.4.role': 'Secrétaire général',
+    'antennes.dauphine.email': 'kryptospheredau@gmail.com',
+    'antennes.dauphine.phone': '+33 6 18 86 28 77',
+    'antennes.dauphine.address': 'Pl. du Maréchal de Lattre de Tassigny, 75016 Paris',
+    'antennes.dauphine.littleDescription': "À Dauphine, KRYPTOSPHERE® connecte économie, finance et technologies émergentes, offrant des ateliers et projets concrets pour former les leaders de demain.",
+
+    // Antenne ESSEC (Paris)
+    'antennes.essec.name': 'ESSEC',
+    'antennes.essec.description': "Une antenne stratégique sur les sujets à l'intersection du Web3, de l'intelligence artificielle et des transformations de la finance d'entreprise et de marché.",
+    'antennes.essec.bureau.1.name': 'Gabriel GERVAIS',
+    'antennes.essec.bureau.1.role': 'Président',
+    'antennes.essec.bureau.2.name': 'Arthur MORVAN',
+    'antennes.essec.bureau.2.role': 'Vice-président',
+    'antennes.essec.bureau.3.name': 'Edgar DELAVOËT',
+    'antennes.essec.bureau.3.role': 'Trésorier',
+    'antennes.essec.bureau.4.name': 'Paul NOËL',
+    'antennes.essec.bureau.4.role': 'Secrétaire général',
+    'antennes.essec.email': 'ksessec@gmail.com',
+    'antennes.essec.phone': '+33 6 52 61 16 79',
+    'antennes.essec.address': '3 avenue Bernard Hirsch CS 50105 95021 Cergy-Pontoise Cedex France',
+    'antennes.essec.littleDescription': "À l'ESSEC, KRYPTOSPHERE® rassemble des étudiants passionnés par la finance, l'innovation et la blockchain, organisant des événements pour explorer les technologies de demain.",
+
+    // Antenne IAE Lyon (Lyon) en français
+    'antennes.lyon.name': 'IAE Lyon',
+    'antennes.lyon.description': 'Chez KRYPTOSPHERE® IAE Lyon, nous connectons la finance, le marketing et le Web3 à travers des projets concrets. Notre objectif : apprendre ensemble, développer nos compétences et préparer les emplois de demain.',
+    'antennes.lyon.bureau.1.name': 'Raphaël PAUTARD',
+    'antennes.lyon.bureau.1.role': 'Président',
+    'antennes.lyon.bureau.2.name': 'Alexandre MANGIN',
+    'antennes.lyon.bureau.2.role': 'Vice-Président',
+    'antennes.lyon.bureau.3.name': 'Jessica DI NOTA',
+    'antennes.lyon.bureau.3.role': 'Trésorière',
+    'antennes.lyon.bureau.4.name': 'Clément CARVALHO',
+    'antennes.lyon.bureau.4.role': 'Secrétaire Général',
+    'antennes.lyon.bureau.5.name': 'Charlotte BECHON',
+    'antennes.lyon.bureau.5.role': 'Responsable des événements',
+    'antennes.lyon.email': 'kryptosphereiaelyon@gmail.com',
+    'antennes.lyon.phone': '+33 7 81 89 20 90',
+    'antennes.lyon.address': '6 Cr Albert Thomas, 69008 Lyon',
+    'antennes.lyon.littleDescription': "KRYPTOSPHERE® IAE Lyon réunit des étudiants de tous horizons pour innover en finance, marketing et Web3, dans un esprit d'équipe et de diversité.",
+
+    // Antenne Télécom SudParis/IMT BS (Paris)
+    'antennes.telecom.name': 'IMTBS/TSP',
+    'antennes.telecom.description': 'Rassemblant des profils ingénieurs et commerciaux, l\'antenne KRYPTOSPHERE® Télécom SudParis/IMT-BS explore l\'IA, le Web3 et la finance, en alliant expertise technique et vision stratégique.',
+    'antennes.telecom.bureau.1.name': 'NEBLE Nicolas',
+    'antennes.telecom.bureau.1.role': 'Président',
+    'antennes.telecom.bureau.2.name': 'DOUMENG Jérémy',
+    'antennes.telecom.bureau.2.role': 'Vice-président',
+    'antennes.telecom.bureau.3.name': 'KENNEDY–MARTINEZ Hugo',
+    'antennes.telecom.bureau.3.role': 'Trésorier',
+    'antennes.telecom.bureau.4.name': 'ABERKANE Samatar',
+    'antennes.telecom.bureau.4.role': 'Secrétaire général',
+    'antennes.telecom.email': 'kryptosphereimt@gmail.com',
+    'antennes.telecom.phone': '+33 6 20 48 10 20',
+    'antennes.telecom.address': '14 Rue Charles Fourier, 91000 Évry-Courcouronnes',
+    'antennes.telecom.littleDescription': "À IMTBS-TSP, KRYPTOSPHERE allie expertise technique et vision stratégique, explorant l'IA, le Web3 et la finance à travers des projets étudiants ambitieux.",
+
+    // Antenne IESEG (Paris)
+    'antennes.ieseg.name': 'IESEG Paris Nanterre',
+    'antennes.ieseg.description': 'Venez rencontrer l\'antenne de l\'IESEG sur le parvis de la Défense, que vous soyez débutant ou expérimenté, pour discuter Web3 et innovation !',
+    'antennes.ieseg.bureau.1.name': 'Guillaume SZONYI',
+    'antennes.ieseg.bureau.1.role': 'Président',
+    'antennes.ieseg.bureau.2.name': 'Nicolas AN SOLERE',
+    'antennes.ieseg.bureau.2.role': 'Vice-Président',
+    'antennes.ieseg.bureau.3.name': 'Adrian Melchior BONNET',
+    'antennes.ieseg.bureau.3.role': 'Pôle Finance',
+    'antennes.ieseg.bureau.4.name': 'Paolo SE',
+    'antennes.ieseg.bureau.4.role': 'Membre',
+    'antennes.ieseg.bureau.5.name': 'Celeste LEBADA',
+    'antennes.ieseg.bureau.5.role': 'Membre',
+    'antennes.ieseg.bureau.6.name': 'Jade LALLOUETTE',
+    'antennes.ieseg.bureau.6.role': 'Membre',
+    'antennes.ieseg.bureau.7.name': 'Louis D\'ANGLADE',
+    'antennes.ieseg.bureau.7.role': 'Membre',
+    'antennes.ieseg.bureau.8.name': 'Clarence COEFFIC',
+    'antennes.ieseg.bureau.8.role': 'Membre',
+    'antennes.ieseg.email': 'kryptosphere.ieseg@gmail.com',
+    'antennes.ieseg.phone': '+33 7 82 76 87 83',
+    'antennes.ieseg.address': '1 parvis de. La Défense - 92044. Paris - La Défense',
+    'antennes.ieseg.littleDescription': "À l'IESEG, KRYPTOSPHERE® accueille tous les curieux sur le parvis de La Défense pour discuter Web3, innovation et entreprendre des initiatives étudiantes.",
+
+    // Pôle AI, DEFI and Financial Engineering
+    'board.poles.aidefi.title': 'Pôle AI, DeFI et Ingénierie Financière',
+    'board.poles.aidefi.description': "Ce pôle a pour mission de créer des partenariats, conférences, workshops et formations à destination des membres désireux de s’investir dans des projets à forte valeur ajoutée. Il anime une communauté d’étudiants ambitieux, désireux de développer une expertise à la fois créative et technique pour concevoir les solutions de demain. En croisant le Web3 avec les technologies et secteurs émergents, il vise des applications concrètes et innovantes.",
+    'board.poles.aidefi.team.member1.name': 'Louis PIALLAT',
+    'board.poles.aidefi.team.member1.role': 'Co-responsable',
+    'board.poles.aidefi.team.member2.name': 'Mark-Killian ZINENBERG',
+    'board.poles.aidefi.team.member2.role': 'Co-responsable',
+
+    // Antenne AES (Paris)
+    'antennes.aes.name': 'Albert Eugenia School',
+    'antennes.aes.description': "À la croisée de l'intelligence artificielle, de la finance et du Web3, notre antenne connecte les étudiants, propulse leurs idées et fait émerger des projets concrets, inclusifs et tournés vers l'impact.",
+    'antennes.aes.bureau.1.name': 'Diane DOUALA',
+    'antennes.aes.bureau.1.role': 'Présidente',
+    'antennes.aes.bureau.2.name': 'Paulin RANDON',
+    'antennes.aes.bureau.2.role': 'Trésorier',
+    'antennes.aes.bureau.3.name': 'Dang VO',
+    'antennes.aes.bureau.3.role': 'Secrétaire général',
+    'antennes.aes.email': ' ',
+    'antennes.aes.phone': ' ',
+    'antennes.aes.address': '18 rue de Paradis, 75010, Paris',
+    'antennes.aes.littleDescription': "À AES, KRYPTOSPHERE® propulse les idées étudiantes et concrétise des projets inclusifs à fort impact, à l'intersection de l'IA, de la finance et du Web3.",
+
+    // Antenne INSEEC Paris
+    'antennes.inseecparis.name': 'INSEEC Paris',
+    'antennes.inseecparis.description': "KRYPTOSPHERE® INSEEC Paris est l'antenne fondée par des étudiants de l'INSEEC MSc, aujourd'hui ouverte à toutes les écoles du groupe Omnes Education à Paris. Notre mission : démocratiser le Web3, l'IA et les technologies émergentes, à travers des formations accessibles, des ateliers pratiques et des événements conviviaux pour apprendre, échanger et se rencontrer.",
+    'antennes.inseecparis.bureau.1.name': 'Antoine MANIVAL',
+    'antennes.inseecparis.bureau.1.role': 'Président',
+    'antennes.inseecparis.bureau.2.name': 'Kylian PERRAULT',
+    'antennes.inseecparis.bureau.2.role': 'Vice-président',
+    'antennes.inseecparis.bureau.3.name': 'Prince TABURY',
+    'antennes.inseecparis.bureau.3.role': 'Trésorier',
+    'antennes.inseecparis.bureau.4.name': 'Khayla RIBESOIS',
+    'antennes.inseecparis.bureau.4.role': 'Secrétaire général',
+    'antennes.inseecparis.bureau.5.name': 'Alexis GRAND',
+    'antennes.inseecparis.bureau.5.role': 'Responsable Partenariats',
+    'antennes.inseecparis.bureau.6.name': 'Selma SANA',
+    'antennes.inseecparis.bureau.6.role': 'Responsable Événementiels',
+    'antennes.inseecparis.bureau.7.name': 'Inass EL BAROUDI',
+    'antennes.inseecparis.bureau.7.role': 'Responsable Marketing/Com',
+    'antennes.inseecparis.email': 'kryptosphere.msc.paris@assos.inseec-france.com',
+    'antennes.inseecparis.phone': '+33 6 71 23 17 78',
+    'antennes.inseecparis.address': '82 Espl. du Général de Gaulle, 92400 Courbevoie',
+    'antennes.inseecparis.littleDescription': "À INSEEC Paris, KRYPTOSPHERE® ouvre le Web3, l'IA et les technologies émergentes à tous les étudiants Omnes, à travers ateliers et événements conviviaux.",
+
+    // Antenne ensIIE Paris
+    'antennes.ensiie.name': 'ensIIE',
+    'antennes.ensiie.description': "L'ensIIE est une école d'ingénieur du groupe Mines-Télécom, spécialisée en informatique et en mathématiques appliquées.\nÀ travers son antenne Kryptosphere, l'école forme et professionnalise ses étudiants aux technologies d'avenir : blockchain, intelligence artificielle, finance, et bien plus encore.\nPortée par une forte culture de l'innovation, l'antenne de l'ENSIIE se distingue par son expertise technique, sa dynamique entrepreneuriale et son engagement dans la transformation numérique.",
+    'antennes.ensiie.bureau.1.name': 'Sarangan UTHAYAN',
+    'antennes.ensiie.bureau.1.role': 'Président',
+    'antennes.ensiie.bureau.2.name': 'Corentin RICHARD',
+    'antennes.ensiie.bureau.2.role': 'Vice-président',
+    'antennes.ensiie.bureau.3.name': 'Tom GILGENKRANTZ',
+    'antennes.ensiie.bureau.3.role': 'Trésorier',
+    'antennes.ensiie.bureau.4.name': 'Lucien LEMOINE',
+    'antennes.ensiie.bureau.4.role': 'Secrétaire général',
+    'antennes.ensiie.email': 'kryptosphere.ensiie@gmail.com',
+    'antennes.ensiie.phone': ' ',
+    'antennes.ensiie.address': '1 place de la Résistance, 91000 Évry-Courcouronnes',
+    'antennes.ensiie.littleDescription': "À l'ensIIE, KRYPTOSPHERE® forme des ingénieurs à la blockchain, l'IA et la finance appliquée, dans une dynamique d'innovation technique.",
+
+    // Antenne NEOMA (Reims, Paris)
+    'antennes.neoma.name': 'NEOMA',
+    'antennes.neoma.description': "L'antenne KRYPTOSPHERE® NEOMA rassemble des étudiants d'école de commerce souhaitant s'éduquer et s'impliquer dans les technologies de notre futur et de notre présent. Notre objectif : démocratiser le web3, la blockchain et la fintech au plus grand nombre, à travers des projets impactants.",
+    'antennes.neoma.bureau.1.name': 'Lucas PAGE',
+    'antennes.neoma.bureau.1.role': 'Président',
+    'antennes.neoma.bureau.2.name': 'Axel KOPEC',
+    'antennes.neoma.bureau.2.role': 'Vice-président',
+    'antennes.neoma.bureau.3.name': 'Emeric PASCAL',
+    'antennes.neoma.bureau.3.role': 'Trésorier',
+    'antennes.neoma.bureau.4.name': 'Hugo DEITSCH',
+    'antennes.neoma.bureau.4.role': 'Vice-Responsable pôle Investment',
+    'antennes.neoma.bureau.5.name': 'Arthur CONTAMINE',
+    'antennes.neoma.bureau.5.role': 'Vice-Responsable pôle Investment',
+    'antennes.neoma.bureau.6.name': 'Victor GOURDOU',
+    'antennes.neoma.bureau.6.role': 'Responsable pôle Business Unit',
+    'antennes.neoma.bureau.7.name': 'Olivier MAZOYER',
+    'antennes.neoma.bureau.7.role': 'Responsable pôle Communication',
+    'antennes.neoma.bureau.8.name': 'Youssef BENSALAH',
+    'antennes.neoma.bureau.8.role': 'Responsable pôle Formation',
+    'antennes.neoma.email': 'lucas.page.22@neoma-bs.com',
+    'antennes.neoma.phone': '+33 6 95 90 28 29',
+    'antennes.neoma.address': '59 Rue Pierre Taittinger, 51100 Reims',
+    'antennes.neoma.littleDescription': "À NEOMA, KRYPTOSPHERE® démocratise la blockchain, la fintech et l'entrepreneuriat, multipliant les initiatives pour former les innovateurs de demain.",
+
+    // Antenne EPITA (Paris)
+    'antennes.epita.name': 'EPITA',
+    'antennes.epita.description': "KRYPTOSPHERE® EPITA est une antenne fondée par des étudiants en apprentissage d'EPITA. Elle est ouverte à l'ensemble des campus EPITA en France ainsi qu'à EPITECH. Notre mission : démocratiser l'intelligence artificielle, l'informatique quantique et les technologies émergentes à travers des formations accessibles, des ateliers pratiques et des événements conviviaux favorisant l'apprentissage et les échanges.",
+    'antennes.epita.bureau.1.name': 'Louise GROSDHOMME',
+    'antennes.epita.bureau.1.role': 'Présidente',
+    'antennes.epita.bureau.2.name': 'Lara HO-CHOUCKFOUNE',
+    'antennes.epita.bureau.2.role': 'Vice-présidente',
+    'antennes.epita.bureau.3.name': 'Clara SION SAUNIER',
+    'antennes.epita.bureau.3.role': 'Trésorière',
+    'antennes.epita.bureau.4.name': 'Ines KACED',
+    'antennes.epita.bureau.4.role': 'Secrétaire générale',
+    'antennes.epita.bureau.5.name': 'Amina SERRANO',
+    'antennes.epita.bureau.5.role': 'Communication',
+    'antennes.epita.email': 'kryptosphere.epita@gmail.com',
+    'antennes.epita.phone': '+33 6 17 20 90 09',
+    'antennes.epita.address': '14-16 Rue Voltaire, 94270 Le Kremlin-Bicêtre',
+    'antennes.epita.littleDescription': "À EPITA, KRYPTOSPHERE® démocratise l'IA, la blockchain et l'innovation auprès des étudiants en informatique, à travers des projets et ateliers pratiques.",
+
+    // Antenne IMT Atlantique (Nantes)
+    'antennes.imtatlantique.name': 'IMT Atlantique',
+    'antennes.imtatlantique.description': "À l'IMT Atlantique, grande école d'ingénieur du Groupe Institut Mines-Télécom, notre antenne KRYPTOSPHERE® rayonne depuis Nantes, Brest et Rennes. Entre océan Atlantique et technologies de rupture, nous explorons ensemble l'IA, le web3, la blockchain et l'ingénierie financière pour construire le monde de demain.",
+    'antennes.imtatlantique.bureau.1.name': 'BADRI Rym',
+    'antennes.imtatlantique.bureau.1.role': 'Présidente',
+    'antennes.imtatlantique.bureau.2.name': 'HADDAD David',
+    'antennes.imtatlantique.bureau.2.role': 'Vice-président',
+    'antennes.imtatlantique.bureau.3.name': 'DE MENEZES Lucas',
+    'antennes.imtatlantique.bureau.3.role': 'Trésorier',
+    'antennes.imtatlantique.bureau.4.name': 'LE NINIVEN Chloé',
+    'antennes.imtatlantique.bureau.4.role': 'Secrétaire général',
+    'antennes.imtatlantique.bureau.5.name': 'OUEDRAOGO David',
+    'antennes.imtatlantique.bureau.5.role': 'Responsable Blockchain',
+    'antennes.imtatlantique.bureau.6.name': 'AUTULY Téodore',
+    'antennes.imtatlantique.bureau.6.role': 'Responsable Blockchain',
+    'antennes.imtatlantique.bureau.7.name': 'ESSALEH Achraf',
+    'antennes.imtatlantique.bureau.7.role': 'Responsable Ingénierie Financière',
+    'antennes.imtatlantique.bureau.8.name': 'OUALDI-DJEBRIL Ines',
+    'antennes.imtatlantique.bureau.8.role': 'Responsable Ingénierie Financière',
+    'antennes.imtatlantique.bureau.9.name': 'DE SAINT SAVIN Thomas',
+    'antennes.imtatlantique.bureau.9.role': 'Responsable IA',
+    'antennes.imtatlantique.bureau.10.name': 'ANDIA Telmo',
+    'antennes.imtatlantique.bureau.10.role': 'Responsable IA',
+    'antennes.imtatlantique.bureau.11.name': 'PAULIAT Gautier',
+    'antennes.imtatlantique.bureau.11.role': 'Responsable Communication',
+    'antennes.imtatlantique.bureau.12.name': 'ELBARI Sara',
+    'antennes.imtatlantique.bureau.12.role': 'Responsable Communication',
+    'antennes.imtatlantique.email': 'kryptosphere.imta@gmail.com',
+    'antennes.imtatlantique.phone': '+33 6 09 60 89 41',
+    'antennes.imtatlantique.address': '4 Rue Alfred Kastler, 44300 Nantes, France',
+    'antennes.imtatlantique.littleDescription': "À IMT Atlantique, KRYPTOSPHERE® rayonne sur l'ouest de la France, explorant l'IA, la blockchain et la finance à travers des projets collaboratifs.",
+
+    // Antenne DEVINCI (Paris)
+    'antennes.devinci.name': 'DeVinci',
+    'antennes.devinci.description': "",
+    'antennes.devinci.bureau.1.name': 'BAUDIN Eden',
+    'antennes.devinci.bureau.1.role': 'Président',
+    'antennes.devinci.bureau.2.name': 'ELGHAZOUI Fawzi',
+    'antennes.devinci.bureau.2.role': 'Vice-président',
+    'antennes.devinci.bureau.3.name': 'THEPAUT Romain',
+    'antennes.devinci.bureau.3.role': 'Vice-président',
+    'antennes.devinci.bureau.4.name': 'WU Kylie',
+    'antennes.devinci.bureau.4.role': 'Trésorière',
+    'antennes.devinci.bureau.5.name': 'CALI Johann',
+    'antennes.devinci.bureau.5.role': 'Secrétaire général',
+    'antennes.devinci.bureau.6.name': 'DEROUIN Alban',
+    'antennes.devinci.bureau.6.role': 'Vice-secrétaire',
+    'antennes.devinci.email': 'kryptosphere@devinci.fr',
+    'antennes.devinci.phone': ' ',
+    'antennes.devinci.address': '32 Rue des 3 Fontanot, 92000 Nanterre',
+    'antennes.devinci.littleDescription': "Au Pôle Léonard de Vinci, KRYPTOSPHERE® réunit ingénieurs, managers et designers autour du Web3, de l'IA et de la blockchain, dans un esprit d'innovation.",
+
+    'antennes.mons.description': ' ',
+    'antennes.mons.littleDescription': "À Mons, KRYPTOSPHERE® favorise l'innovation technologique et la finance, créant un hub étudiant dynamique au cœur de la Belgique.",
+    'antennes.hult.littleDescription': "À Hult Business School London, KRYPTOSPHERE® connecte des étudiants internationaux autour de la finance traditionnelle, des crypto, de l'innovation et de l'entrepreneuriat.",
+    'antennes.escp.littleDescription': "À l'ESCP, KRYPTOSPHERE® se distingue par ses projets innovants en crypto, IA et finance, dans un environnement académique cosmopolite et dynamique.",
+
+    
+    'donation.btcModal.title': 'Donner en Bitcoin via Tor',
+    'donation.btcModal.body': "💡 Pour un don en BTC, le lien qui suit (xxxx.onion) doit être ouvert avec le navigateur Tor Browser (don en BTC sans intermédiaires, directement via notre noeud Bitcoin).\n\nSi vous n'avez pas ce navigateur (Tor Browser), vous pouvez le télécharger ici : https://www.torproject.org/\n\nSinon, copiez ce lien et ouvrez-le dans TOR :\n2l4udju3groh7qbt2lvy7iy3x3rg2n5ajmikzqttdcomd2lyxjb3c6ad.onion/apps/2R7Z45MZWJUfxtVdF55Yz1eJx9Sn/pos",
+    'donation.btcModal.button': 'Ouvrir le lien dans Tor',
+    'donation.btcModal.close': 'Fermer'
+  },
+  en: {
+    // Navigation
+    'nav.home': 'Home',
+    'nav.board': 'Board',
+    'nav.association': 'Association',
+    'nav.branches': 'Branches',
+    'nav.events': 'Events',
+    'nav.publications': 'Publications',
+    'nav.donate': 'Donate',
+
+    // Hero
+    'hero.title': 'The first student',
+    'hero.title2': 'Tech',
+    'hero.title3': 'network',
+    'hero.title4': 'in Europe',
+    'hero.subtitle': 'The physical network of the digital world: crypto-assets, AI and cybersecurity',
+    'hero.join': 'Join the association',
+    'hero.branches': 'See our branches',  
+
+    // Stats
+    'stats.title': 'KRYPTOSPHERE® IN NUMBERS',
+    'stats.subtitle': 'The leading EU student network about Tech and cryptocurrencies since 2017',
+    'stats.students': 'Students',
+    'stats.universities': 'Universities',
+    'stats.events': 'Events',
+    'stats.branches': 'Branches',
+
+    // Sponsors
+    'sponsors.title': 'They support us',
+    'sponsors.sponsors': 'Sponsors',
+    'sponsors.partners': 'Partners',
+    'sponsors.universities': 'Universities',
+    'sponsors.media': 'They talk about us',
+
+    // Publications
+    'publications.title': 'Our latest publications',
+    'publications.subtitle': 'Discover our latest blockchain analyses and research',
+    'publications.readMore': 'Read more',
+    'publications.viewAll': 'View all',
+
+    // Pages
+    'board.title': 'Our Board',
+    'board.subtitle': 'Meet the passionate individuals who keep KRYPTOSPHERE® running every day. Diverse profiles and complementary teams, united by their trust and passion for cryptocurrencies since 2017.',
+    'board.teamTag': 'The team that runs and manages KRYPTOSPHERE®',
+    'board.year': 'Year:',
+    'board.member1.name': 'Raphaël PAUTARD',
+    'board.member1.role': 'President',
+    'board.member1.description': 'Family Officer specialized in crypto & co-founder of theintern.fun',
+    'board.member1.school': 'IAE Lyon - KRYPTOSPHERE® FRANCE',
+    'board.member2.name': 'Antoine MANIVAL',
+    'board.member2.role': 'Vice President', 
+    'board.member2.description': 'Growth & Marketing @NFT Paris, Partnerships @Ledger.',
+    'board.member2.school': 'INSEEC - KRYPTOSPHERE® FRANCE',
+    'board.member3.name': 'Mathieu LARUELLE',
+    'board.member3.role': 'Treasurer',
+    'board.member3.description': 'Blockchain & Software Engineer @Cometh & Louis.finance, Ex Treasurer of KRYPTOSPHERE® DeVinci',
+    'board.member3.school': 'ESILV - KRYPTOSPHERE® FRANCE',
+    'board.member4.name': 'Louis PIALLAT',
+    'board.member4.role': 'Secretary',
+    'board.member4.description': 'Ex VP Kryptosphere ESSEC. Student in finance at ESSEC and Maths at Sorbonne;',
+    'board.member4.school': 'ESSEC - KRYPTOSPHERE® FRANCE',
+    'board.member5.name': 'Thibault LANGLOIS-BERTHELOT',
+    'board.member5.role': 'Chairman & Founder',
+    'board.member5.description': 'Trainee Lawyer & Law PhD',
+    'board.member5.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member6.name': 'William DERENNE',
+    'board.member6.role': 'Admin & Founding Member', 
+    'board.member6.description': 'Solidity professor, Ethereum specialist',  
+    'board.member6.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member7.name': 'Serge PETROFF',
+    'board.member7.role': 'Chairman',
+    'board.member7.description': 'Consultant specializing in decentralized finance',
+    'board.member7.school': 'RENNES SB',
+    'board.member8.name': 'Guillaume GIRARD',
+    'board.member8.role': 'Admin & Founding Member',
+    'board.member8.description': 'Consultant specializing in Bitcoin mining',
+    'board.member8.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member9.name': 'Othmane EL MEJJADE',
+    'board.member9.role': 'Admin & Delegate Treasurer',
+    'board.member9.description': 'Consultant and specialist in international marketing',
+    'board.member9.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member10.name': 'Fanis MICHALAKIS',
+    'board.member10.role': 'Chairman & Founder',
+    'board.member10.description': 'Developer specializing in Bitcoin & Lightning Network at LN MARKETS',
+    'board.member10.school': 'KRYPTOSPHERE® CENTRALE MARSEILLE',
+    'board.member11.name': 'Titouan GALPIN',
+    'board.member11.role': 'Delegate Vice-President',
+    'board.member11.description': 'Founder @Luphy.io',
+    'board.member11.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member12.name': 'Julie MARTIN',
+    'board.member12.role': 'Communication Manager',
+    'board.member12.description': 'Product Management & AI consultant',
+    'board.member12.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member13.name': 'Manon SERAUT',
+    'board.member13.role': 'Training Manager',
+    'board.member13.description': 'Human resources manager @ARCHIMED',
+    'board.member13.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member14.name': 'Harmonie LEFORT',
+    'board.member14.role': 'Partnerships Manager',
+    'board.member14.description': 'Relationship Manager @CACEIS',
+    'board.member14.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member15.name': 'Marie CECCHINI',
+    'board.member15.role': 'Partnerships Manager',
+    'board.member15.description': 'Portfolio manager in private wealth management @Amplegest',
+    'board.member15.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member16.name': 'Aymane TABET',
+    'board.member16.role': 'R&D Manager',
+    'board.member16.description': 'Risk Advisory consultant @DELOITTE',
+    'board.member16.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member17.name': 'Karolina GORNA',
+    'board.member17.role': 'Delegate Chairman',
+    'board.member17.description': 'PhD engineer student @LEDGER',
+    'board.member17.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member18.name': 'Pierre LOURDELET',
+    'board.member18.role': 'Delegate Vice President',
+    'board.member18.description': 'Data Scientist at Mirakl',
+    'board.member18.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member19.name': 'Mathias VALLET',  
+    'board.member19.role': 'Delegate Chairman',
+    'board.member19.description': 'Co-founder and President @RUFIJI',
+    'board.member19.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member20.name': 'Gabriel BEGAZO',
+    'board.member20.role': 'Delegate Treasurer',
+    'board.member20.description': 'Co-founder and President of DUSA LABS',
+    'board.member20.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member21.name': 'Maxime DIENGER',
+    'board.member21.role': 'Delegate President',
+    'board.member21.description': 'Dev Advocate at Ripple',
+    'board.member21.school': 'EMLV',
+    'board.member22.name': 'Léo JOUANNY',
+    'board.member22.role': 'Delegate Vice President',
+    'board.member22.description': 'Founder and CEO @Dusa Labs',
+    'board.member22.school': 'EMLV',
+    'board.member23.name': 'Louis FARCOT',
+    'board.member23.role': 'Delegate Treasurer',
+    'board.member23.description': 'Researcher specializing in blockchains',
+    'board.member23.school': 'EMLV',
+    'board.member24.name': 'Mathias BOUGON',
+    'board.member24.role': 'Delegate President',
+    'board.member24.description': 'Associate Business Developer at HubSpot',
+    'board.member24.school': 'EMLV',
+    'board.member25.name': 'Rayan THIBAULT',
+    'board.member25.role': 'Delegate Vice President',
+    'board.member25.description': 'Project Engineer - Technical Management at agap2',
+    'board.member25.school': 'EMLV',
+    'board.member26.name': 'Gildas SOEIRO',
+    'board.member26.role': 'Delegate Secretary',
+    'board.member26.description': 'Program Manager KRYPTOSPHERE® Accelerator',
+    'board.member26.school': 'EMLV',
+    'board.member27.name': 'Thibault LANGLOIS-BERTHELOT',
+    'board.member27.role': 'Representative & Founder',
+    'board.member27.description': 'Legal representative supporting our NPO in its complex aspects.',
+    'board.member27.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member29.name': 'Brice BONNET',
+    'board.member29.role': 'Training Manager, Writing/Analysis',
+    'board.member29.description': 'Internal Auditor specializing in Banking and Finance.',
+    'board.member29.school': 'KRYPTOSPHERE® RENNES',
+    'board.member31.name': 'Jules FOA',
+    'board.member31.role': 'Chairman & Founder',
+    'board.member31.description': 'Web3 advisor',
+    'board.member31.school': 'KRYPTOSPHERE® ACCELERATOR',
+    'board.member36.name': 'Hadrien JAUBERT',
+    'board.member36.role': 'Chairman',
+    'board.member36.description': 'Developer @Google',
+    'board.member36.school': 'KRYPTOSPHERE® CENTRALE MARSEILLE',
+    'board.member55.name': 'Thomas BENETREAU',
+    'board.member55.role': 'Event Manager',
+    'board.member55.description': 'M&A @Wavestone',
+    'board.member55.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member56.name': 'Clément BIDAN',
+    'board.member56.role': 'Active Member',
+    'board.member56.description': 'Software Engineer @Datadog',
+    'board.member56.school': 'KRYPTOSPHERE® CENTRALE MARSEILLE',
+    'board.member57.name': 'Jérémie PONTIS',
+    'board.member57.role': 'Treasurer',
+    'board.member57.description': 'Senior Consultant @Talan Consulting',
+    'board.member57.school': 'KRYPTOSPHERE® IMTBS-TSP Paris',
+    'board.member28.name': 'Choi JANGHYUN',
+    'board.member28.role': 'Head of IoT',
+    'board.member28.description': "Ph.D Student @KAIST",
+    'board.member28.school': 'KRYPTOSPHERE® IMTBS-TSP',
+    'board.member34.name': 'Aymeric LABEDAN',
+    'board.member34.role': 'Partnership Manager',
+    'board.member34.description': "CTO Advisory Consultant @Wavestone",
+    'board.member34.school': 'KRYPTOSPHERE® IMTBS-TSP',
+    'board.member35.name': 'Morgan RICHAUD',
+    'board.member35.role': 'Head of Research',
+    'board.member35.description': "Marketing & Design Specialist @Deskoin",
+    'board.member35.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member39.name': 'Mathis GONTIER DELAUNAY',
+    'board.member39.role': 'Secrétaire Général délégué',
+    'board.member39.description': "Co-founder and Head of Research @Morpho",
+    'board.member39.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member40.name': 'Paul FRAMBOT',
+    'board.member40.role': 'Active Member',
+    'board.member40.description': "Founder @Morpho",
+    'board.member40.school': 'KRYPTOSPHERE® IMTBS-TSP Paris',
+    'board.member44.name': 'Andy COUSINEAU',
+    'board.member44.role': 'Chairman & Founder',
+    'board.member44.description': "Blockchain engineer @SAHAR",
+    'board.member44.school': 'KRYPTOSPHERE® ESEO ANGERS',
+    'board.member52.name': 'Paul WOLF',
+    'board.member52.role': 'Active Member',
+    'board.member52.description': "Crypto Lead @Mastercard",
+    'board.member52.school': 'KRYPTOSPHERE® ESSEC',
+    'board.member53.name': 'Pierre-Antoine ARSAGUET',
+    'board.member53.role': 'Head of Artificial Intelligence',
+    'board.member53.description': "Quantitative Researcher @QRT",
+    'board.member53.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member54.name': 'Imen SACI',
+    'board.member54.role': 'Partnership Lead',
+    'board.member54.description': "Blockchain & Innovation Consultant @IBM",
+    'board.member54.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member41.name': 'Louis DOIDY',
+    'board.member41.role': 'Training Lead',
+    'board.member41.description': "Data Scientist @LVMH",
+    'board.member41.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member42.name': 'Nandy Bâ',
+    'board.member42.role': 'Head of Research & DeFI',
+    'board.member42.description': "Web3 advisor",
+    'board.member42.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member43.name': 'Nour RAMMAL',
+    'board.member43.role': 'Chairman',
+    'board.member43.description': "Student in Information Technology and Maths @Polytechnique Institute",
+    'board.member43.school': 'KRYPTOSPHERE® IMTBS-TSP',
+    'board.member48.name': 'Pierre-Louis HABOURY',
+    'board.member48.role': 'Head of blockchain programming',
+    'board.member48.description': "Co-Founder & CTO @Dusa Labs",
+    'board.member48.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member49.name': 'Lilyan BASTIEN',
+    'board.member49.role': 'Active Member',
+    'board.member49.description': "Blockchain engineer @RA2 TECH",
+    'board.member49.school': 'KRYPTOSPHERE® DEVINCI',
+    'board.member50.name': 'Pierre FONTAINE',
+    'board.member50.role': 'General Secretary',
+    'board.member50.description': "Business Analyst @Accenture",
+    'board.member50.school': 'KRYPTOSPHERE® IMTBS-TSP',
+    'board.member33.name': 'Maxence LAISNE',
+    'board.member33.role': 'Delegate General Secretary',
+    'board.member33.description': "Tech, Product, Analytics @AMAZON",
+    'board.member33.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member37.name': 'Benoît PREAU',
+    'board.member37.role': 'Head of Writing',
+    'board.member37.description': "Consultant Energy Transition @Capgemini",
+    'board.member37.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member46.name': 'Manon COIFFARD',
+    'board.member46.role': 'Event Manager',
+    'board.member46.description': "Consultant ERP @mc2i",
+    'board.member46.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member47.name': 'Quentin BAILLY',
+    'board.member47.role': 'Active Member',
+    'board.member47.description': "M&A @UBS",
+    'board.member47.school': 'KRYPTOSPHERE® LYON',
+    'board.member51.name': 'Sara FEORE',
+    'board.member51.role': 'Chairman',
+    'board.member51.description': "Co-founder @CompasSup",
+    'board.member51.school': 'KRYPTOSPHERE® Science Po',
+    'board.member30.name': 'Charles-André GOICHOT',
+    'board.member30.role': 'Delegate Treasurer',
+    'board.member30.description': "Developer",
+    'board.member30.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member32.name': 'Maksym ANDRUSHKIV',
+    'board.member32.role': 'Head of Communication',
+    'board.member32.description': "Digital Transformation @Staizen Europe",
+    'board.member32.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member38.name': 'Emile BREMOND',
+    'board.member38.role': 'Head of Redaction',
+    'board.member38.description': "Writer @CRYPTOAST",
+    'board.member38.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member45.name': 'Nasreddine GRIHMA',
+    'board.member45.role': 'Head of Research',
+    'board.member45.description': "Blockchain advisor @IBM",
+    'board.member45.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member59.name': 'Lery PHELIPPEAU',
+    'board.member59.role': 'Vice-President & Co-founder',
+    'board.member59.description': 'Marketing @CoinJar',
+    'board.member59.school': 'KRYPTOSPHERE® LONDON',
+    'board.member60.name': 'Thibaut GRAVELLE-VIVIEN',
+    'board.member60.role': 'Chairman & Founder',
+    'board.member60.description': 'Growth Associate @Merkle Science',
+    'board.member60.school': 'KRYPTOSPHERE® LONDON',
+    'board.member61.name': 'Jean-Baptiste BAROU',
+    'board.member61.role': 'Head of Ambassadors / Alumni',
+    'board.member61.description': 'Outsourcing Analyst @Euronext',
+    'board.member61.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member62.name': 'Mark-Killian ZINENBERG',
+    'board.member62.role': 'DeFi Lead',
+    'board.member62.description': 'Quantitative Developer Intern @Natixis ',
+    'board.member62.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member63.name': 'Salim LAKHAL',
+    'board.member63.role': 'Chairman',
+    'board.member63.description': 'Student @Télécom SudParis',
+    'board.member63.school': 'KRYPTOSPHERE® IMTBS-TSP Paris',
+    'board.member64.name': 'Maxime LACROIX',
+    'board.member64.role': 'Tech Advisor',
+    'board.member64.description': 'Team Lead & Platform Engineer @InstaDeep',
+    'board.member64.school': 'KRYPTOSPHERE® FRANCE',
+    'board.member65.name': 'Emmanuel MARQUETTE',
+    'board.member65.role': 'Chairman & Founder',
+    'board.member65.description': 'Strategy Manager @Lombard Odier',
+    'board.member65.school': 'KRYPTOSPHERE® IMTBS-TSP Paris',
+
+    'board.former.president': 'Former President',
+    'board.former.description': 'Former president who led our network of associations to new horizons since 2017.',
+    'board.former.description2023': 'Exceptional leadership during the year 2023.',
+
+    'board.polesTitle': 'Poles',
+    'board.poles.partnership.title': 'Partnership Pole',
+    'board.poles.partnership.description': 'The partnership pole is responsible for acquiring and maintaining external relationships with the association\'s partners and sponsors. Their responsibilities are the spokespersons of the association\'s visions and contribute to its visibility in the public space.',
+    'board.poles.partnership.team.member1.name': 'Jules-Grégoire Lafond',
+    'board.poles.partnership.team.member1.role': 'Manager',
+    'board.poles.partnership.team.member1.description': 'Co-Founder & Ex-President of Kryptosphere NEOMA. Sales at HSBC',
+    'board.poles.partnership.team.member2.name': 'Tevfik Erdal',
+    'board.poles.partnership.team.member2.role': 'Co-manager',
+    'board.poles.partnership.team.member2.description': 'Ex-Vice President of Kryptosphere NEOMA. Customer Relationship at UBS',
+    'board.poles.partnership.team.member3.name': 'Rayan Kaabache',
+    'board.poles.partnership.team.member3.role': 'Co-Manager',
+    'board.poles.partnership.team.member3.description': 'Co-Founder & Ex-President of Kryptosphere Jussieu. Cybersecurity student at Sorbonne University',
+    'board.poles.communication.title': 'Communication & Marketing Pole',
+    'board.poles.communication.description': ['The Communication & Marketing Pole embodies the image of KRYPTOSPHERE and ensures its visibility among the general public as a key player in the Web3 ecosystem. Through creative and well-managed communication strategies, it promotes our events, initiatives and the engagement of our members.',
+     'Social media, branding (visual identity), content creation, visual construction, media partnerships or even digital campaigns: our mission is to shape and disseminate the identity of KRYPTOSPHERE to make it a key player in the European Web3 ecosystem.'],
+    'board.poles.communication.team.member1.name': 'Alexandre MANGIN',
+    'board.poles.communication.team.member1.role': 'Manager',
+    'board.poles.communication.team.member1.description': 'VP KRYPTOSPHERE iaelyon, Marketing manager at iExec, Community Manager at The Big Whale',
+    'board.poles.communication.team.member2.name': 'Alexandra AH-SOUNE',
+    'board.poles.communication.team.member2.role': 'Co-manager',
+    'board.poles.communication.team.member2.description': 'Ex-Communication Manager of Kryptosphere ESSEC. Private Banking at Société Générale, student in Finance',
+
+    'association.title': 'The Association',
+    'association.subtitle': 'Since September 2017, KRYPTOSPHERE® has established itself as the European reference in the student crypto ecosystem. Discover our history, values and vision for the future.',
+    'association.historyTag': 'Our history and mission',
+    'association.mission': 'Our Mission',
+    'association.missionText': 'Democratize access to crypto-assets and train the next generation in new digital technologies.',
+    'association.values': 'Our Values',
+    'association.valuesText': 'Trust, knowledge sharing, collaboration and academic excellence in service of  our volunteer network.',
+    'association.vision': 'Our Vision',
+    'association.visionText': 'Bridge the gap between innovation in AI, blockchain, quantum computing and higher education in Europe.',
+    'association.journey': 'Our Journey',
+
+    'branches.title': 'Our Branches',
+    'branches.subtitle': 'Discover our branches across Europe and join the Kryptosphere community closest to you.',
+    'branches.discover': 'Discover the branch',
+    'branches.localOffice': 'Local office',
+    'branches.presenceTag': 'Our presence in Europe',
+    'branches.parisDescription': 'Our Parisian branch, at the heart of the French blockchain ecosystem.',
+    'branches.parisDescription2': "Discover KRYPTOSPHERE DEVINCI, one of our most active Parisian branches!",
+    'branches.parisDescription3': "Discover KRYPTOSPHERE IMTBS-TSP, one of our oldest and most active Parisian branches!",
+    'branches.nantesDescription': 'The Nantes branch, dynamic and innovative in western France.',
+    'branches.lilleDescription': 'The Lille branch, close to the Belgian border and very active.',
+    'branches.londonDescription': 'Our first international branch, in the heart of London\'s City.',
+    'branches.parisDescription4': "The ESCP branch, created in April 2022, as prestigious as it is competent. Specialization: crypto, AI.",
+    'branches.lyon': "Created in February 2024, KS IAE LYON shines through the annual organization of our flagship event: the KryptoTour!",
+    'branches.lyonDescription': "Created in February 2024, KS IAE LYON shines through the annual organization of our flagship event: the KryptoTour!",
+    'branches.monsDescription': 'The Mons branch, close to the European institutions in Brussels.',
+
+    'branches.backToList': 'Back to branches',
+    'branches.bureauYear': '{city} Office 2025',
+    'branches.contact': 'Contact us',
+    'branches.gallery': 'Gallery',
+    'branches.location': 'Location',
+    'branches.locationMap': 'Location map',
+
+    'events.title': 'Events',
+    'events.subtitle': 'Join our exclusive events and dive into the blockchain universe. Workshops, conferences and networking with crypto ecosystem leaders.',
+    'events.agendaTag': 'Event agenda',
+    'events.lumaButton': 'See all our events on Lu.ma',
+    'events.upcoming': 'Upcoming events',
+    'events.past': 'Past events',
+
+    'publications.page.title': 'Publications',
+    'publications.page.subtitle': 'Discover our analyses, guides and reports on crypto-assets, blockchain and other technologies. Resources created by and for the student community.',
+    'publications.resourcesTag': 'Resources and educational content',
+    'publications.all': 'All',
+    'publications.kryptopaper': 'KryptoPaper',
+    'publications.articles': 'Articles',
+    'publications.reports': 'Reports',
+    'publications.studies': 'Studies',
+    'publications.read': 'Read',
+    'publications.pdf': 'PDF',
+    'publications.newsletter.title': 'Stay informed',
+    'publications.newsletter.subtitle': 'Receive our latest publications directly in your inbox.',
+    'publications.newsletter.placeholder': 'Your email',
+    'publications.newsletter.subscribe': 'Subscribe',
+
+    'donation.title': 'Make a donation',
+    'donation.subtitle': 'Your support helps us democratize access to blockchain technologies and train the next generation of European innovators.',
+    'donation.supportTag': 'Support our mission',
+    'donation.impact': 'Your impact',
+    'donation.training': 'Train students',
+    'donation.trainingDesc': 'Funding workshops, hackathons and blockchain training',
+    'donation.development': 'Develop our branches',
+    'donation.developmentDesc': 'Expansion to new institutions and countries',
+    'donation.events': 'Fund events',
+    'donation.eventsDesc': 'Organization of free conferences and meetups',
+    'donation.choose': 'Donate by card',
+    'donation.choose2': 'Donate in Bitcoin',
+    'donation.amount': 'Donation amount',
+    'donation.customAmount': 'Custom amount',
+    'donation.paymentMethod': 'Payment method',
+    'donation.card': 'Credit card',
+    'donation.crypto': 'Cryptocurrency',
+    'donation.summary': 'Summary',
+    'donation.amountLabel': 'Donation amount:',
+    'donation.info': 'Your donation will be used to support our educational activities and community development.',
+    'donation.finalize': 'Finalize donation',
+    'donation.security': 'Your donations are secure and used exclusively for association activities.',
+    'donation.transparency': 'Financial transparency',
+    'donation.transparencyCRYPTO': 'Donate in cryptocurrency',
+    'donation.transparencyBTC': 'Our Bitcoin node',
+    'donation.transparencyText': 'We publish part of our financial information to ensure transparency about our activities.',
+    'donation.transparencyTextBTC': 'We have been voluntarily running our own Bitcoin node for several years: support us!',
+    'donation.transparencyTextCRYPTO': 'We have been accepting direct donations in ETH or BTC since 2021.',
+    'donation.reports': 'View our official data',
+    'donation.reportsBTCNODE': 'Discover our Bitcoin node',
+    'donation.reportsBTC': 'Donate in BTC (via TOR Browser)',
+    'donation.reportsETH': 'Donate in ETH',
+
+    // Antenne Details
+    'antenne.team': 'Team',
+    'antenne.president': 'President',
+    'antenne.vicePresident': 'Vice-President',
+    'antenne.secretary': 'Secretary',
+    'antenne.treasurer': 'Treasurer',
+    'antenne.contact': 'Contact',
+    'antenne.gallery': 'Gallery',
+    'antenne.location': 'Location',
+    'antenne.email': 'Email',
+    'antenne.phone': 'Phone',
+    'antenne.address': 'Address',
+    'antenne.notFound': 'Branch not found',
+
+    // Countries
+    'countries.france': 'France',
+    'countries.uk': 'United Kingdom',
+    'countries.belgium': 'Belgium',
+
+    // Cities
+    'cities.paris': 'Paris',
+    'cities.nantes': 'Nantes',
+    'cities.lille': 'Lille',
+    'cities.london': 'London',
+    'cities.brussels': 'Brussels',
+    'cities.reims': 'Reims',
+    'cities.lyon': 'Lyon',
+    'cities.mons': 'Mons',
+
+    // Team roles
+    'roles.president': 'President',
+    'roles.vicePresident': 'Vice-President',
+    'roles.secretary': 'Secretary',
+    'roles.treasurer': 'Treasurer',
+    'roles.presidente': 'President',
+    'roles.vicePresidente': 'Vice-President',
+    'roles.tresoriere': 'Treasurer',
+
+    // Footer Legal
+    'footer.legal.disclaimer': 'Disclaimer',
+    'footer.legal.privacy': 'Privacy Policy',
+    'footer.legal.terms': 'Terms of Use',
+    'footer.legal.cookies': 'Cookie Policy',
+    'footer.legal.legal': 'Legal Notice',
+    'footer.description': 'The reference student association in the Tech universe. Join us to explore the technologies of tomorrow.',
+    'footer.navigation': 'Navigation',
+    'footer.copyright': '2017 — 2025 | KRYPTOSPHERE FRANCE | All rights reserved.',
+
+    // Contact
+    'contact.email': 'Email',
+    'contact.phone': 'Phone',
+    'contact.address': 'Address',
+
+    // Timeline years
+    'timeline.2017.title': 'Launch of KRYPTOSPHERE®',
+    'timeline.2017.description': 'Foundation of our network by 4 students passionate about Bitcoin',
+    'timeline.2017.achievements': ['First branch at KEDGE Marseille', '10 founding members of the network', 'First articles on Bitcoin'], 
+    'timeline.2018.title': 'National expansion and creation of KRYPTOSPHERE® FRANCE',
+    'timeline.2018.description': 'Opening of the first branches in the Paris region and partnerships with several prestigious schools and universities',
+    'timeline.2018.achievements': ['4 active branches', '50+ members', 'First blockchain hackathon'],
+    'timeline.2019.title': "Development and diversification in AI",
+    'timeline.2019.description': "Creation of new branches to support the development of our network and diversification in AI",
+    'timeline.2019.achievements': ['6 active branches', '100+ members', 'First KryptoPaper'],
+    'timeline.2020.title': 'Deployment of our communication strategy',
+    'timeline.2020.description': 'Adaptation to new communication modes',
+    'timeline.2020.achievements': ['Virtual events', '10+ active branches', '200+ members'],
+    'timeline.2021.title': 'Diversification in cybersecurity',
+    'timeline.2021.description': "Opening of the first branches within computer science schools",
+    'timeline.2021.achievements': ['New sponsors', 'Ongoing expansion', '300+ members'],
+    'timeline.2022.title': 'Educational innovation',
+    'timeline.2022.description': 'Launch of advanced internal training programs on blockchain and crypto-assets',
+    'timeline.2022.achievements': ['Certified internal training', '15+ branches', '400+ members'],
+    'timeline.2023.title': 'Consolidation of our network',
+    'timeline.2023.description': "Recognition as a reference in France in higher education",
+    'timeline.2023.achievements': ['500+ members', '20+ branches', '3 countries'],
+    'timeline.2024.title': 'Internationalization in London and Belgium',
+    'timeline.2024.description': 'Creation of two KRYPTOSPHERE® associations in London and Belgium',
+    'timeline.2024.achievements': ['London', 'Belgium', '700+ members'],
+    'timeline.2025.title': 'New horizons',
+    'timeline.2025.description': "Targeting new emerging technologies and consolidating our position as a leader in European higher education",
+    'timeline.2025.achievements': ['Artificial Intelligence', 'Quantum Computers', 'Cybersecurity', '800+ members'],
+
+    // Publication items
+    'publication.kryptopaper22.title': 'KryptoPaper Edition 22',
+    'publication.kryptopaper22.date': 'MARCH 2022',
+    'publication.kryptopaper22.description': 'Analysis of crypto trends in 2022 and future perspectives',
+    'publication.kryptopaper22.readTime': '15 min',
+    'publication.defi.title': 'KryptoPaper Edition 2021',
+    'publication.defi.date': 'MARCH 2021',
+    'publication.defi.description': 'Comprehensive introduction to crypto-assets',  
+    'publication.defi.readTime': '10 min',
+    'publication.report2024.title': 'Impact Report 2024',
+    'publication.report2024.date': 'FEBRUARY 2024',
+    'publication.report2024.description': 'Quantitative assessment of our network\'s impact on the crypto market',
+    'publication.report2024.readTime': '15 min',
+    'publication.kryptopaper24.title': 'KryptoPaper Edition 22',
+    'publication.kryptopaper24.date': 'APRIL 2022',
+    'publication.kryptopaper24.description': 'News, DeFi, Crypto-assets, In-depth articles',
+    'publication.kryptopaper24.readTime': '18 min',
+    'publication.research.title': 'PhD Thesis',
+    'publication.research.date': 'AUGUST 2023',
+    'publication.research.description': 'Blockchain in the light of law and identity',
+    'publication.research.readTime': '60+ min',
+    'publication.smartcontracts.title': 'Archived Articles',
+    'publication.smartcontracts.date': '2018 - 2024',
+    'publication.smartcontracts.description': 'Find archives of our old articles to follow multiple technological evolutions',
+    'publication.smartcontracts.readTime': '+60 min',
+
+    // General
+    'language': 'en',
+
+    // Antenne Dauphine (Paris)
+    'antennes.dauphine.name': 'Dauphine',
+    'antennes.dauphine.description': 'Because the future is built today, the Dauphine branch of Kryptosphere introduces curious minds to the technologies transforming our world.',
+    'antennes.dauphine.bureau.1.name': 'Ray MOUAWAD',
+    'antennes.dauphine.bureau.1.role': 'President',
+    'antennes.dauphine.bureau.2.name': 'Yanis BOUKHARI',
+    'antennes.dauphine.bureau.2.role': 'Vice President',
+    'antennes.dauphine.bureau.3.name': 'Emiel VAN RENTERGHEM',
+    'antennes.dauphine.bureau.3.role': 'Treasurer',
+    'antennes.dauphine.bureau.4.name': 'Amine MELLITI',
+    'antennes.dauphine.bureau.4.role': 'General Secretary',
+    'antennes.dauphine.email': 'kryptospheredau@gmail.com',
+    'antennes.dauphine.phone': '+33 6 18 86 28 77',
+    'antennes.dauphine.address': 'Pl. du Maréchal de Lattre de Tassigny, 75016 Paris',
+    'antennes.dauphine.littleDescription': "At Dauphine, KRYPTOSPHERE connects economics, finance, and emerging technologies, offering workshops and real-world projects to train tomorrow's leaders.",
+
+    // Antenne ESSEC (Paris)
+    'antennes.essec.name': 'ESSEC',
+    'antennes.essec.description': 'The ESSEC branch of KRYPTOSPHERE® is a dynamic community of students passionate about blockchain, AI and cybersecurity. We organize events, workshops and conferences to explore the latest trends in these fields.',
+    'antennes.essec.bureau.1.name': 'Alexandre MANGIN',
+    'antennes.essec.bureau.1.role': 'President',
+    'antennes.essec.bureau.2.name': 'Louis PIALLAT',
+    'antennes.essec.bureau.2.role': 'Vice President',
+    'antennes.essec.bureau.3.name': 'Thibault LANGLOIS-BERTHELOT',
+    'antennes.essec.bureau.3.role': 'Treasurer',
+    'antennes.essec.bureau.4.name': 'Louis FARCOT',
+    'antennes.essec.bureau.4.role': 'Delegate Secretary',
+    'antennes.essec.bureau.5.name': 'Othmane EL MEJJADE',
+    'antennes.essec.bureau.5.role': 'Communication Manager',
+    'antennes.essec.email': 'ksessec@gmail.com',
+    'antennes.essec.phone': '+33 6 12 34 5678',
+    'antennes.essec.address': '3 Avenue Bernard Hirsch, 95000 Cergy',
+    'antennes.essec.littleDescription': "At ESSEC, KRYPTOSPHERE brings together students passionate about finance, innovation, and blockchain, organizing events to explore tomorrow's technologies.",
+
+    // Antenne IAE Lyon (Lyon)
+    'antennes.lyon.name': 'IAE Lyon',
+    'antennes.lyon.description': 'At KRYPTOSPHERE IAE Lyon, we connect finance, marketing, and Web3 through real-world projects. Our goal: learn together, develop our skills, and prepare for the jobs of tomorrow.',
+    'antennes.lyon.bureau.1.name': 'Raphaël PAUTARD',
+    'antennes.lyon.bureau.1.role': 'President',
+    'antennes.lyon.bureau.2.name': 'Alexandre MANGIN',
+    'antennes.lyon.bureau.2.role': 'Vice President',
+    'antennes.lyon.bureau.3.name': 'Jessica DI NOTA',
+    'antennes.lyon.bureau.3.role': 'Treasurer',
+    'antennes.lyon.bureau.4.name': 'Clément CARVALHO',
+    'antennes.lyon.bureau.4.role': 'General Secretary',
+    'antennes.lyon.bureau.5.name': 'Charlotte BECHON',
+    'antennes.lyon.bureau.5.role': 'Events Manager',
+    'antennes.lyon.email': 'kryptosphereiaelyon@gmail.com',
+    'antennes.lyon.phone': '+33 7 81 89 20 90',
+    'antennes.lyon.address': '6 Cr Albert Thomas, 69008 Lyon',
+    'antennes.lyon.littleDescription': "KRYPTOSPHERE IAE Lyon brings together students from all backgrounds to innovate in finance, marketing, and Web3, fostering teamwork and diversity.",
+
+    // Antenne Télécom SudParis/IMT BS (Paris)
+    'antennes.telecom.name': 'IMTBS/TSP',
+    'antennes.telecom.description': 'Bringing together engineering and business profiles, the KRYPTOSPHERE Télécom SudParis/IMT-BS branch explores AI, Web3, and finance, combining technical expertise with strategic vision.',
+    'antennes.telecom.bureau.1.name': 'NEBLE Nicolas',
+    'antennes.telecom.bureau.1.role': 'President',
+    'antennes.telecom.bureau.2.name': 'DOUMENG Jérémy',
+    'antennes.telecom.bureau.2.role': 'Vice President',
+    'antennes.telecom.bureau.3.name': 'KENNEDY–MARTINEZ Hugo',
+    'antennes.telecom.bureau.3.role': 'Treasurer',
+    'antennes.telecom.bureau.4.name': 'ABERKANE Samatar',
+    'antennes.telecom.bureau.4.role': 'General Secretary',
+    'antennes.telecom.email': 'kryptosphereimt@gmail.com',
+    'antennes.telecom.phone': '+33 6 20 48 10 20',
+    'antennes.telecom.address': '14 Rue Charles Fourier, 91000 Évry-Courcouronnes',
+    'antennes.telecom.littleDescription': "At IMTBS-TSP, KRYPTOSPHERE combines technical expertise and strategic vision, exploring AI, Web3, and finance through ambitious student projects.",
+
+    // Antenne IESEG (Paris)
+    'antennes.ieseg.name': 'IESEG Paris Nanterre',
+    'antennes.ieseg.description': 'Come meet the IESEG branch on the Défense esplanade, whether you\'re a beginner or experienced, to discuss Web3 and innovation!',
+    'antennes.ieseg.bureau.1.name': 'Guillaume SZONYI',
+    'antennes.ieseg.bureau.1.role': 'President',
+    'antennes.ieseg.bureau.2.name': 'Nicolas An SOLERE',
+    'antennes.ieseg.bureau.2.role': 'Vice-President',
+    'antennes.ieseg.bureau.3.name': 'Adrian Melchior BONNET',
+    'antennes.ieseg.bureau.3.role': 'Finance Pole',
+    'antennes.ieseg.bureau.4.name': 'Paolo Se',
+    'antennes.ieseg.bureau.4.role': 'Membre',
+    'antennes.ieseg.bureau.5.name': 'Celeste LEBADA',
+    'antennes.ieseg.bureau.5.role': 'Member',
+    'antennes.ieseg.bureau.6.name': 'Jade LALLOUETTE',
+    'antennes.ieseg.bureau.6.role': 'Member',
+    'antennes.ieseg.bureau.7.name': 'Louis D\'ANGLADE',
+    'antennes.ieseg.bureau.7.role': 'Member',
+    'antennes.ieseg.bureau.8.name': 'Clarence COEFFIC',
+    'antennes.ieseg.bureau.8.role': 'Member',
+    'antennes.ieseg.email': 'kryptosphere.ieseg@gmail.com',
+    'antennes.ieseg.phone': '+33 7 82 76 87 83',
+    'antennes.ieseg.address': '1 parvis de. La Défense - 92044. Paris - La Défense',
+    'antennes.ieseg.littleDescription': "At IESEG, KRYPTOSPHERE welcomes all curious minds at La Défense to discuss Web3, innovation, and launch student initiatives.",
+
+    // AI, DEFI and Financial Engineering Pole
+    'board.poles.aidefi.title': 'AI, DEFI and Financial Engineering Pole',
+    'board.poles.aidefi.description': "This pole was created to foster partnerships, conferences, workshops, and training for members wishing to immerse themselves in high value-added missions in innovative fields. The goal? To animate a community of ambitious students seeking to build expertise that allows them to be both competent and creative in creating tomorrow's solutions. Thus, we evolve Web3 to correlate it with trending technologies and sectors to find real applications.",
+    'board.poles.aidefi.team.member1.name': 'Louis PIALLAT',
+    'board.poles.aidefi.team.member1.role': 'Co-lead',
+    'board.poles.aidefi.team.member2.name': 'Mark-Killian ZINENBERG',
+    'board.poles.aidefi.team.member2.role': 'Co-lead',
+
+    // Antenne AES (Paris)
+    'antennes.aes.name': 'Albert Eugenia School',
+    'antennes.aes.description': "At the crossroads of artificial intelligence, finance, and Web3, our branch connects students, propels their ideas, and brings to life concrete, inclusive, and impact-driven projects.",
+    'antennes.aes.bureau.1.name': 'Diane DOUALA',
+    'antennes.aes.bureau.1.role': 'President',
+    'antennes.aes.bureau.2.name': 'Paulin RANDON',
+    'antennes.aes.bureau.2.role': 'Treasurer',
+    'antennes.aes.bureau.3.name': 'Dang VO',
+    'antennes.aes.bureau.3.role': 'General Secretary',
+    'antennes.aes.email': ' ',
+    'antennes.aes.phone': ' ',
+    'antennes.aes.address': '18 rue de Paradis, 75010, Paris',
+    'antennes.aes.littleDescription': "At AES, KRYPTOSPHERE propels student ideas and brings to life inclusive, high-impact projects at the intersection of AI, finance, and Web3.",
+
+    // Antenne INSEEC Paris
+    'antennes.inseecparis.name': 'INSEEC Paris',
+    'antennes.inseecparis.description': "Kryptosphere INSEEC Paris is the branch founded by INSEEC MSc students, now open to all schools of the Omnes Education group in Paris. Our mission: to democratize Web3, AI, and emerging technologies through accessible training, practical workshops, and friendly events to learn, exchange, and meet.",
+    'antennes.inseecparis.bureau.1.name': 'Antoine MANIVAL',
+    'antennes.inseecparis.bureau.1.role': 'President',
+    'antennes.inseecparis.bureau.2.name': 'Kylian PERRAULT',
+    'antennes.inseecparis.bureau.2.role': 'Vice-President',
+    'antennes.inseecparis.bureau.3.name': 'Prince TABURY',
+    'antennes.inseecparis.bureau.3.role': 'Treasurer',
+    'antennes.inseecparis.bureau.4.name': 'Khayla RIBESOIS',
+    'antennes.inseecparis.bureau.4.role': 'General Secretary',
+    'antennes.inseecparis.bureau.5.name': 'Alexis GRAND',
+    'antennes.inseecparis.bureau.5.role': 'Partnerships Manager',
+    'antennes.inseecparis.bureau.6.name': 'Selma SANA',
+    'antennes.inseecparis.bureau.6.role': 'Events Manager',
+    'antennes.inseecparis.bureau.7.name': 'Inass EL BAROUDI',
+    'antennes.inseecparis.bureau.7.role': 'Marketing/Com Manager',
+    'antennes.inseecparis.email': 'kryptosphere.msc.paris@assos.inseec-france.com',
+    'antennes.inseecparis.phone': '+33 6 71 23 17 78',
+    'antennes.inseecparis.address': '82 Espl. du Général de Gaulle, 92400 Courbevoie',
+    'antennes.inseecparis.littleDescription': "At INSEEC Paris, KRYPTOSPHERE opens Web3, AI, and emerging tech to all Omnes students, through accessible workshops and friendly events.",
+
+    // Antenne ensIIE Paris
+    'antennes.ensiie.name': 'ensIIE',
+    'antennes.ensiie.description': "ensIIE is an engineering school of the Mines-Télécom group, specialized in computer science and applied mathematics.\nThrough its Kryptosphere branch, the school trains and professionalizes its students in future technologies: blockchain, artificial intelligence, finance, and much more.\nDriven by a strong culture of innovation, the ENSIIE branch stands out for its technical expertise, entrepreneurial drive, and commitment to digital transformation.",
+    'antennes.ensiie.bureau.1.name': 'Sarangan UTHAYAN',
+    'antennes.ensiie.bureau.1.role': 'President',
+    'antennes.ensiie.bureau.2.name': 'Corentin RICHARD',
+    'antennes.ensiie.bureau.2.role': 'Vice President',
+    'antennes.ensiie.bureau.3.name': 'Tom GILGENKRANTZ',
+    'antennes.ensiie.bureau.3.role': 'Treasurer',
+    'antennes.ensiie.bureau.4.name': 'Lucien LEMOINE',
+    'antennes.ensiie.bureau.4.role': 'General Secretary',
+    'antennes.ensiie.email': 'kryptosphere.ensiie@gmail.com',
+    'antennes.ensiie.phone': ' ',
+    'antennes.ensiie.address': '1 place de la Résistance, 91000 Évry-Courcouronnes',
+    'antennes.ensiie.littleDescription': "At ensIIE, KRYPTOSPHERE trains engineers in blockchain, AI, and applied finance, fostering a culture of technical innovation.",
+
+    // Antenne NEOMA (Reims, Paris)
+    'antennes.neoma.name': 'NEOMA',
+    'antennes.neoma.description': "The KRYPTOSPHERE NEOMA branch brings together business school students wishing to learn about and get involved in the technologies of our future and present. Our goal: to democratize web3, blockchain, and fintech to as many people as possible through impactful projects.",
+    'antennes.neoma.bureau.1.name': 'Lucas PAGE',
+    'antennes.neoma.bureau.1.role': 'President',
+    'antennes.neoma.bureau.2.name': 'Axel KOPEC',
+    'antennes.neoma.bureau.2.role': 'Vice President',
+    'antennes.neoma.bureau.3.name': 'Emeric PASCAL',
+    'antennes.neoma.bureau.3.role': 'Treasurer',
+    'antennes.neoma.bureau.4.name': 'Hugo DEITSCH',
+    'antennes.neoma.bureau.4.role': 'Deputy Head of Investment',
+    'antennes.neoma.bureau.5.name': 'Arthur CONTAMINE',
+    'antennes.neoma.bureau.5.role': 'Deputy Head of Investment',
+    'antennes.neoma.bureau.6.name': 'Victor GOURDOU',
+    'antennes.neoma.bureau.6.role': 'Head of Business Unit',
+    'antennes.neoma.bureau.7.name': 'Olivier MAZOYER',
+    'antennes.neoma.bureau.7.role': 'Head of Communication',
+    'antennes.neoma.bureau.8.name': 'Youssef BENSALAH',
+    'antennes.neoma.bureau.8.role': 'Head of Training',
+    'antennes.neoma.email': 'lucas.page.22@neoma-bs.com',
+    'antennes.neoma.phone': '+33695902829',
+    'antennes.neoma.address': '59 Rue Pierre Taittinger, 51100 Reims',
+    'antennes.neoma.littleDescription': "At NEOMA, KRYPTOSPHERE democratizes blockchain, fintech, and entrepreneurship, multiplying initiatives to train tomorrow's innovators.",
+
+    // Antenne EPITA (Paris)
+    'antennes.epita.name': 'EPITA',
+    'antennes.epita.description': "Kryptosphere EPITA is a branch founded by EPITA apprenticeship students. It is open to all EPITA campuses in France as well as EPITECH. Our mission: democratize artificial intelligence, quantum computing, and emerging technologies through accessible training, hands-on workshops, and friendly events fostering learning and exchange.",
+    'antennes.epita.bureau.1.name': 'Louise GROSDHOMME',
+    'antennes.epita.bureau.1.role': 'President',
+    'antennes.epita.bureau.2.name': 'Lara HO-CHOUCKFOUNE',
+    'antennes.epita.bureau.2.role': 'Vice President',
+    'antennes.epita.bureau.3.name': 'Clara SION SAUNIER',
+    'antennes.epita.bureau.3.role': 'Treasurer',
+    'antennes.epita.bureau.4.name': 'Ines KACED',
+    'antennes.epita.bureau.4.role': 'General Secretary',
+    'antennes.epita.bureau.5.name': 'Amina SERRANO',
+    'antennes.epita.bureau.5.role': 'Communication',
+    'antennes.epita.email': 'kryptosphere.epita@gmail.com',
+    'antennes.epita.phone': '+33 6 17 20 90 09',
+    'antennes.epita.address': '14-16 Rue Voltaire, 94270 Le Kremlin-Bicêtre',
+    'antennes.epita.littleDescription': "At EPITA, KRYPTOSPHERE brings AI, blockchain, and innovation to computer science students, through practical projects and workshops.",
+
+    // Antenne IMT Atlantique (Nantes)
+    'antennes.imtatlantique.name': 'IMT Atlantique',
+    'antennes.imtatlantique.description': "At IMT Atlantique, a prestigious engineering school of the Institut Mines-Télécom Group, our Kryptosphere branch radiates from Nantes, Brest, and Rennes. Between the Atlantic Ocean and breakthrough technologies, we explore together AI, web3, blockchain, and financial engineering to build tomorrow's world.",
+    'antennes.imtatlantique.bureau.1.name': 'BADRI Rym',
+    'antennes.imtatlantique.bureau.1.role': 'President',
+    'antennes.imtatlantique.bureau.2.name': 'HADDAD David',
+    'antennes.imtatlantique.bureau.2.role': 'Vice President',
+    'antennes.imtatlantique.bureau.3.name': 'DE MENEZES Lucas',
+    'antennes.imtatlantique.bureau.3.role': 'Treasurer',
+    'antennes.imtatlantique.bureau.4.name': 'LE NINIVEN Chloé',
+    'antennes.imtatlantique.bureau.4.role': 'General Secretary',
+    'antennes.imtatlantique.bureau.5.name': 'OUEDRAOGO David',
+    'antennes.imtatlantique.bureau.5.role': 'Blockchain Manager',
+    'antennes.imtatlantique.bureau.6.name': 'AUTULY Téodore',
+    'antennes.imtatlantique.bureau.6.role': 'Blockchain Manager',
+    'antennes.imtatlantique.bureau.7.name': 'ESSALEH Achraf',
+    'antennes.imtatlantique.bureau.7.role': 'Financial Engineering Manager',
+    'antennes.imtatlantique.bureau.8.name': 'OUALDI-DJEBRIL Ines',
+    'antennes.imtatlantique.bureau.8.role': 'Financial Engineering Manager',
+    'antennes.imtatlantique.bureau.9.name': 'DE SAINT SAVIN Thomas',
+    'antennes.imtatlantique.bureau.9.role': 'AI Manager',
+    'antennes.imtatlantique.bureau.10.name': 'ANDIA Telmo',
+    'antennes.imtatlantique.bureau.10.role': 'AI Manager',
+    'antennes.imtatlantique.bureau.11.name': 'PAULIAT Gautier',
+    'antennes.imtatlantique.bureau.11.role': 'Communication Manager',
+    'antennes.imtatlantique.bureau.12.name': 'ELBARI Sara',
+    'antennes.imtatlantique.bureau.12.role': 'Communication Manager',
+    'antennes.imtatlantique.email': 'kryptosphere.imta@gmail.com',
+    'antennes.imtatlantique.phone': '+33 6 09 60 89 41',
+    'antennes.imtatlantique.address': '4 Rue Alfred Kastler, 44300 Nantes, France',
+    'antennes.imtatlantique.littleDescription': "At IMT Atlantique, KRYPTOSPHERE shines across western France, exploring AI, blockchain, and finance through collaborative projects.",
+
+    // Antenne DEVINCI (Paris)
+    'antennes.devinci.name': 'DeVinci',
+    'antennes.devinci.description': "",
+    'antennes.devinci.bureau.1.name': 'BAUDIN Eden',
+    'antennes.devinci.bureau.1.role': 'President',
+    'antennes.devinci.bureau.2.name': 'ELGHAZOUI Fawzi',
+    'antennes.devinci.bureau.2.role': 'Vice President',
+    'antennes.devinci.bureau.3.name': 'THEPAUT Romain',
+    'antennes.devinci.bureau.3.role': 'Vice President',
+    'antennes.devinci.bureau.4.name': 'WU Kylie',
+    'antennes.devinci.bureau.4.role': 'Treasurer',
+    'antennes.devinci.bureau.5.name': 'CALI Johann',
+    'antennes.devinci.bureau.5.role': 'General Secretary',
+    'antennes.devinci.bureau.6.name': 'DEROUIN Alban',
+    'antennes.devinci.bureau.6.role': 'Deputy Secretary',
+    'antennes.devinci.email': 'kryptosphere@devinci.fr',
+    'antennes.devinci.phone': ' ',
+    'antennes.devinci.address': '32 Rue des 3 Fontanot, 92000 Nanterre',
+    'antennes.devinci.littleDescription': "At Pôle Léonard de Vinci, KRYPTOSPHERE unites engineers, managers, and designers around Web3, AI, and blockchain, in a spirit of innovation.",
+
+    'antennes.mons.description': ' ',
+    'antennes.mons.littleDescription': "In Mons, KRYPTOSPHERE fosters tech innovation and finance, creating a dynamic student hub in the heart of Belgium.",
+    'antennes.hult.littleDescription': "At Hult Business School London, KRYPTOSPHERE connects international students around traditional finance, crypto, innovation, and global entrepreneurship.",
+    'antennes.escp.littleDescription': "At ESCP, KRYPTOSPHERE stands out for innovative projects in crypto, AI, and finance, within a dynamic and cosmopolitan academic environment.",
+
+    // Antenne Hult (London)
+    'antennes.london.name': 'Hult Business School',
+    'antennes.london.description': "KRYPTOSPHERE® LONDON embodies our European expansion, connecting traditional finance and the crypto universe.",
+
+
+    
+    'donation.btcModal.title': 'Donate in Bitcoin via Tor',
+    'donation.btcModal.body': "💡 For a BTC donation, the following link (xxxx.onion) must be opened with the Tor Browser (BTC donation without intermediaries, directly via our Bitcoin node).\n\nIf you don't have this browser (Tor Browser), you can download it here: https://www.torproject.org/\n\nOtherwise, copy this link and open it in TOR:\n2l4udju3groh7qbt2lvy7iy3x3rg2n5ajmikzqttdcomd2lyxjb3c6ad.onion/apps/2R7Z45MZWJUfxtVdF55Yz1eJx9Sn/pos",
+    'donation.btcModal.button': 'Open link in Tor',
+    'donation.btcModal.close': 'Close'
+  }
+};
